consolidated and excluded test files before hosting
