--- canvas.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Plotly Scroll Zoom Test</title>
-    <!-- Include Plotly library -->
-    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-</head>
-<body>
-    <div id="myDiv" style="width: 600px; height: 400px;"></div>
-
-    <script>
-var trace1 = {
-
-y: [5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5],
-
-mode: 'markers',
-
-marker: {
-
-  size: 40,
-
-  color: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39]
-
-}
-
-};
-
-
-var data = [trace1];
-
-
-var layout = {
-
-title: 'Scatter Plot with a Color Dimension'
-
-};
-
-
-Plotly.newPlot('myDiv', data, layout, {scrollZoom: true});    </script>
-</body>
-</html>
