--- parseWBBFile.js
@@ -1,31 +1,0 @@
-// Define a function to parse a Wii Balance Board (WBB) file
-function parseWbbFile(fileAddress) {
-    const time = [];
-    const signal = [];
-    const fs = require('fs'); // Assuming Node.js for file operations
-    const data = fs.readFileSync(fileAddress, 'utf8').split('\n');
-
-    for (const line of data) {
-        if (line.length > 0) {
-            const parts = line.split(' ');
-            const t = 0.001 * parseFloat(parts[0]);
-            const x = parseFloat(parts[1]);
-            const y = parseFloat(parts[2]);
-            time.push(t);
-            signal.push([x, y]);
-        }
-    }
-
-    return [time, signal];
-}
-
-if (true) { // Replace 'true' with your condition or environment check
-    // Create an instance of the SWARII class
-    const resamplingMethod = new SWARII({ windowSize: 0.25, desiredFrequency: 25 });
-
-    // Parse the WBB file
-    const [time, signal] = parseWbbFile('example_wbb_statokinesigram.txt');
-
-    // Resample the signal using SWARII
-    const [resampledTime, resampledSignal] = resamplingMethod.resample(time, signal);
-}
