--- swarii.js
@@ -1,1941 +1,0 @@
-resample(data) { // Accept an array of arrays where each inner array contains [time, x, y] data
-    const windowSize = 1; // Set the window size to 1
-    const desiredFrequency = 25; // Set the desired frequency to 25
-    const currentTime = Math.max(0, data[0][0]);
-    const outputData = [];
-
-    while (currentTime < data[data.length - 1][0]) {
-        // Create an array to store relevant data point indices within the window
-        const relevantTimes = [];
-
-        // Find data points that fall within the current sliding window
-        for (let t = 0; t < data.length; t++) {
-            if (Math.abs(data[t][0] - currentTime) < windowSize * 0.5) {
-                relevantTimes.push(t);
-            }
-        }
-
-        if (relevantTimes.length === 1) {
-            // If there's only one relevant data point, use it as is
-            const t = relevantTimes[0];
-            outputData.push([currentTime, data[t][1], data[t][2]]);
-        } else {
-            let sumX = 0;
-            let sumY = 0;
-            let weightSum = 0;
-
-            // Calculate the weighted average of data points within the window
-            for (let i = 0; i < relevantTimes.length; i++) {
-                const t = relevantTimes[i];
-                let leftBorder, rightBorder;
-
-                if (i === 0 || t === 0) {
-                    leftBorder = Math.max(data[0][0], currentTime - windowSize * 0.5);
-                } else {
-                    leftBorder = 0.5 * (data[t][0] + data[t - 1][0]);
-                }
-
-                if (i === relevantTimes.length - 1) {
-                    rightBorder = Math.min(data[data.length - 1][0], currentTime + windowSize * 0.5);
-                } else {
-                    rightBorder = 0.5 * (data[t + 1][0] + data[t][0]);
-                }
-
-                const w = rightBorder - leftBorder;
-                sumX += data[t][1] * w; // Accumulate weighted X values
-                sumY += data[t][2] * w; // Accumulate weighted Y values
-                weightSum += w;
-            }
-
-            const resampledX = sumX / weightSum;
-            const resampledY = sumY / weightSum;
-
-            outputData.push([currentTime, resampledX, resampledY]);
-        }
-
-        // Move the sliding window by incrementing the current time
-        currentTime += 1 / desiredFrequency;
-    }
-
-    return outputData; // Return resampled data as an array of arrays
-}
-
-// let jaSignal = [
-//     [0.017, -0.0145824659674, -0.381828569452],
-//     [0.021, -0.0148970945298, -0.343829201286],
-//     [0.05, 0.0101892602241, -0.291166396417],
-//     [0.068, -0.0760989735418, -0.270120519792],
-//     [0.07200000000000001, -0.0506985391088, -0.266308292712],
-//     [0.092, -0.050478259921, -0.292304495523],
-//     [0.109, -0.053628617426, -0.309515747609],
-//     [0.127, -0.0375896154209, -0.32601629595],
-//     [0.132, -0.0473934228147, -0.317384606604],
-//     [0.149, -0.0441718793429, -0.319338931859],
-//     [0.153, -0.0659652076355, -0.397547312709],
-//     [0.181, -0.0976631894742, -0.430917573287],
-//     [0.183, -0.138811178382, -0.474661852268],
-//     [0.211, -0.148575750694, -0.50285455296],
-//     [0.214, -0.103179221698, -0.573884568961],
-//     [0.242, -0.160403982773, -0.607898178605],
-//     [0.26, -0.172897423142, -0.61106818487],
-//     [0.265, -0.0964156643531, -0.599768664572],
-//     [0.292, -0.109242893123, -0.584650773209],
-//     [0.295, -0.0836609544667, -0.604357504839],
-//     [0.323, -0.0354884123216, -0.598970067433],
-//     [0.341, -0.0292527041582, -0.606512180297],
-//     [0.34600000000000003, -0.0262930610648, -0.602953960569],
-//     [0.373, -0.0131654492673, -0.622260302965],
-//     [0.376, 0.00293102248677, -0.632445432853],
-//     [0.403, 0.0127814122232, -0.650608461822],
-//     [0.421, -0.00650822810825, -0.642577098893],
-//     [0.427, 0.0382542179403, -0.676884349938],
-//     [0.454, 0.0157450728486, -0.640839516954],
-//     [0.458, 0.0413695416735, -0.63257481599],
-//     [0.486, 0.0440921143184, -0.60075896186],
-//     [0.489, 0.0732911489432, -0.600626859754],
-//     [0.518, 0.0762582917838, -0.615583073331],
-//     [0.534, 0.0636136106761, -0.657024897577],
-//     [0.539, 0.0798710576472, -0.677950545221],
-//     [0.5670000000000001, 0.064413164193, -0.745756246225],
-//     [0.5700000000000001, 0.0453957098311, -0.76451775783],
-//     [0.599, 0.0201583512584, -0.793826565982],
-//     [0.619, 0.0457418586532, -0.816464239486],
-//     [0.621, 0.0620902450751, -0.842680856676],
-//     [0.648, 0.0399611035196, -0.855350954058],
-//     [0.652, 0.0302422074977, -0.850295899071],
-//     [0.681, 0.040045695486, -0.871284465949],
-//     [0.6980000000000001, 0.0275750430984, -0.886975571023],
-//     [0.7020000000000001, 0.0148387625456, -0.92570825077],
-//     [0.729, 0.0211060937106, -0.928049926083],
-//     [0.733, -0.026418623869, -0.956959939564],
-//     [0.761, 0.0182714517456, -0.967209314593],
-//     [0.778, -0.00382502526526, -0.988653561726],
-//     [0.783, 0.034260056852, -0.978024108376],
-//     [0.811, 0.00246535630267, -0.977351552053],
-//     [0.8130000000000001, -0.00414239584057, -0.97125690751],
-//     [0.841, -0.0201254909958, -0.953193538795],
-//     [0.858, -0.010674877353, -0.939569557457],
-//     [0.864, -0.052106199848, -0.931051913961],
-//     [0.892, 0.00174351553595, -0.882751686201],
-//     [0.9, -0.055823163777, -0.865624147147],
-//     [0.918, -0.0914671271004, -0.835078895582],
-//     [0.936, -0.0722029416282, -0.826793190167],
-//     [0.9530000000000001, -0.0657912362388, -0.828500372913],
-//     [0.971, -0.0754407252987, -0.812138185423],
-//     [0.977, -0.0818490327061, -0.801052583066],
-//     [1.004, -0.0951395814667, -0.736549287653],
-//     [1.008, -0.127741992697, -0.676657251484],
-//     [1.036, -0.0963426848743, -0.608900668082],
-//     [1.053, -0.0900711110752, -0.585665578978],
-//     [1.058, -0.0999117547339, -0.549300806249],
-//     [1.085, -0.0870516082638, -0.54216564997],
-//     [1.089, -0.0934835688751, -0.553277841541],
-//     [1.117, -0.0806295948173, -0.549734378357],
-//     [1.119, -0.0838834692315, -0.566973366653],
-//     [1.137, -0.0967510805811, -0.559007788434],
-//     [1.141, -0.102898388257, -0.561871265615],
-//     [1.168, -0.0902205098397, -0.550234214585],
-//     [1.171, -0.090526867343, -0.536192001941],
-//     [1.198, -0.157551229337, -0.534029434016],
-//     [1.202, -0.131756346826, -0.538088236521],
-//     [1.222, -0.0874387575762, -0.50505383713],
-//     [1.249, -0.164075587714, -0.550916343868],
-//     [1.252, -0.125333601816, -0.565261597571],
-//     [1.2790000000000001, -0.112593612036, -0.558107694931],
-//     [1.283, -0.127927312338, -0.596161756225],
-//     [1.311, -0.153797564255, -0.592858546193],
-//     [1.332, -0.150105065715, -0.612175835511],
-//     [1.333, -0.114874698513, -0.64377491051],
-//     [1.361, -0.127618186769, -0.619954467734],
-//     [1.364, -0.124599819101, -0.612249926468],
-//     [1.3920000000000001, -0.131443664729, -0.594123477561],
-//     [1.3960000000000001, -0.128360367138, -0.551053127555],
-//     [1.413, -0.109092125937, -0.562991188193],
-//     [1.426, -0.135290184273, -0.533530464725],
-//     [1.438, -0.0839964935951, -0.528513766954],
-//     [1.448, -0.103412843715, -0.536494510046],
-//     [1.476, -0.0776777380755, -0.537122726564],
-//     [1.477, -0.090085834426, -0.550434075892],
-//     [1.504, -0.115773261322, -0.547293666925],
-//     [1.522, -0.0776375249829, -0.505238380148],
-//     [1.528, -0.0745023023847, -0.512834734191],
-//     [1.556, -0.090907105645, -0.457715003217],
-//     [1.559, -0.0715891793493, -0.476989518938],
-//     [1.587, -0.0810909777576, -0.477666263655],
-//     [1.604, -0.0678824550244, -0.507929006349],
-//     [1.609, -0.0836140154971, -0.539168360468],
-//     [1.6380000000000001, -0.070617274685, -0.564128194657],
-//     [1.639, -0.0419114696462, -0.579144555511],
-//     [1.667, -0.105544274425, -0.597823421433],
-//     [1.684, -0.0385203100491, -0.628189147145],
-//     [1.69, -0.0953349634433, -0.675940949369],
-//     [1.717, -0.0888647200744, -0.70214672474],
-//     [1.721, -0.0950490623096, -0.741046186886],
-//     [1.748, -0.101561060201, -0.732753434999],
-//     [1.766, -0.0792733727444, -0.69973099202],
-//     [1.7710000000000001, -0.0888647200744, -0.713457868329],
-//     [1.798, -0.11119083777, -0.719630551989],
-//     [1.802, -0.06635677873, -0.71382267057],
-//     [1.828, -0.139721853899, -0.736679461209],
-//     [1.846, -0.0979034645794, -0.77658901648],
-//     [1.852, -0.107133152886, -0.822619795992],
-//     [1.8800000000000001, -0.144657686893, -0.850349266692],
-//     [1.883, -0.147854457535, -0.859600790004],
-//     [1.933, -0.154075831371, -0.867398869112],
-//     [1.944, -0.116377808032, -0.862905867301],
-//     [1.945, -0.138141056431, -0.892311737622],
-//     [1.959, -0.147637083412, -0.891118389202],
-//     [1.964, -0.15081646225, -0.909635461055],
-//     [1.992, -0.128415171047, -0.931181155074],
-//     [1.994, -0.182681334034, -0.976440994859],
-//     [2.023, -0.176094129723, -1.00609754416],
-//     [2.041, -0.17289829577, -1.02737466715],
-//     [2.0460000000000003, -0.155941167739, -1.04703931894],
-//     [2.073, -0.17268228723, -1.02891421023],
-//     [2.077, -0.169355169625, -0.98792883781],
-//     [2.104, -0.179274973479, -0.959436613562],
-//     [2.122, -0.163878644538, -0.90143785738],
-//     [2.136, -0.0745410438531, -0.890946989434],
-//     [2.153, -0.148759884111, -0.826344080081],
-//     [2.157, -0.0787162925144, -0.789722135483],
-//     [2.184, -0.06617016908, -0.755440335817],
-//     [2.202, -0.0375963661038, -0.715674275038],
-//     [2.207, -0.0534752025542, -0.696299366822],
-//     [2.234, -0.0504821114832, -0.668333833934],
-//     [2.238, -0.0825089657243, -0.653145790975],
-//     [2.266, -0.0379733363172, -0.622488384474],
-//     [2.283, -0.0988051846633, -0.613362877071],
-//     [2.289, -0.0925538803595, -0.617532972506],
-//     [2.317, -0.0767934152972, -0.624317684955],
-//     [2.321, -0.0575832801384, -0.631116846106],
-//     [2.349, -0.127432555183, -0.668065372963],
-//     [2.367, -0.120891213997, -0.6869585596],
-//     [2.371, -0.0788681548061, -0.72570240862],
-//     [2.399, -0.0916461507732, -0.725198773856],
-//     [2.4010000000000002, -0.0887634647615, -0.697943553518],
-//     [2.428, -0.133666366736, -0.698458241438],
-//     [2.436, -0.130457033299, -0.723100841073],
-//     [2.453, -0.139732451776, -0.786492780209],
-//     [2.481, -0.120071357266, -0.828432132131],
-//     [2.483, -0.147872116765, -0.879229453027],
-//     [2.512, -0.102974084714, -0.906026265317],
-//     [2.5140000000000002, -0.141097899919, -0.935748450655],
-//     [2.5420000000000003, -0.115889298103, -0.918987703621],
-//     [2.559, -0.0779183816928, -0.916997524638],
-//     [2.564, -0.0296378407788, -0.935489073312],
-//     [2.592, -0.0325453804798, -0.97009658896],
-//     [2.596, -0.0261166662209, -0.985227389937],
-//     [2.616, -0.0708022919827, -0.993905043884],
-//     [2.6430000000000002, -0.0126845618793, -1.04731651565],
-//     [2.646, -0.0542119087112, -1.04514462069],
-//     [2.673, 0.0382766475198, -1.02703244014],
-//     [2.676, -0.00952828305339, -1.04495368145],
-//     [2.704, 0.0064760585657, -1.04136926504],
-//     [2.722, -0.00323431916425, -1.03965659875],
-//     [2.727, 0.0571231538376, -0.9969330435],
-//     [2.754, 0.044397988069, -0.982745359563],
-//     [2.757, 0.0534195270975, -0.928207115323],
-//     [2.7840000000000003, 0.0599495143799, -0.947375937455],
-//     [2.802, 0.0984041846755, -0.930933234403],
-//     [2.8080000000000003, 0.0793045304815, -0.945712812712],
-//     [2.836, 0.127113350421, -0.947858192885],
-//     [2.838, 0.066460236168, -0.956411145972],
-//     [2.866, 0.0919522023556, -0.964567182567],
-//     [2.883, 0.120372763527, -0.968644944981],
-//     [2.8890000000000002, 0.127085970542, -0.985522789479],
-//     [2.9170000000000003, 0.152844706408, -0.971756232107],
-//     [2.919, 0.162659512222, -0.968353025493],
-//     [2.948, 0.19771018837, -0.930330706404],
-//     [2.966, 0.178858328533, -0.938741903265],
-//     [2.971, 0.181845219534, -0.913980102583],
-//     [2.998, 0.19691653247, -0.907754533002],
-//     [3.001, 0.19001147773, -0.89474354568],
-//     [3.028, 0.189862547545, -0.919667362479],
-//     [3.032, 0.114082308294, -0.938519339856],
-//     [3.059, 0.14295276706, -0.943567391272],
-//     [3.077, 0.0731227989085, -0.968681531093],
-//     [3.083, 0.136891676842, -0.975798652291],
-//     [3.111, 0.12418359228, -0.98213615171],
-//     [3.113, 0.108395866193, -1.01309800226],
-//     [3.141, 0.117755757386, -0.98558375837],
-//     [3.144, 0.114550183965, -0.990178822168],
-//     [3.162, 0.127519822944, -0.998383717454],
-//     [3.164, 0.130739051203, -0.992296524682],
-//     [3.192, 0.146592190788, -0.983650933832],
-//     [3.196, 0.143469822065, -0.992926373252],
-//     [3.223, 0.137641520968, -1.02880070889],
-//     [3.227, 0.0769388800861, -1.05298229402],
-//     [3.254, 0.102808318338, -1.07235040306],
-//     [3.2720000000000002, 0.0741144082502, -1.1015267503],
-//     [3.277, 0.0997318199114, -1.13691641975],
-//     [3.306, 0.0487996334597, -1.13916235204],
-//     [3.3080000000000003, 0.0395309743027, -1.14133285441],
-//     [3.336, 0.00749926075795, -1.14016045676],
-//     [3.339, 0.0265120147526, -1.10619414181],
-//     [3.367, 0.0393529334039, -1.09071443306],
-//     [3.384, -0.0251828833934, -1.0609168142],
-//     [3.3890000000000002, -0.0120861528456, -1.08610739255],
-//     [3.4170000000000003, 0.0100321204197, -1.07030239086],
-//     [3.4210000000000003, -0.0477426154729, -1.05434263457],
-//     [3.4490000000000003, -0.0637133811917, -1.02329047793],
-//     [3.466, -0.0478358237746, -1.02286649276],
-//     [3.471, -0.0861620483208, -1.00735968773],
-//     [3.498, -0.0640109244775, -0.990588834941],
-//     [3.5020000000000002, -0.0835443708054, -0.948122024283],
-//     [3.5300000000000002, -0.112629720797, -0.882494457261],
-//     [3.547, -0.0907885203198, -0.798096931635],
-//     [3.553, -0.113499710404, -0.757796516959],
-//     [3.5820000000000003, -0.107593269279, -0.714614658954],
-//     [3.583, -0.0791886396539, -0.693540663601],
-//     [3.611, -0.0983289390001, -0.689938738051],
-//     [3.615, -0.124029225976, -0.690741134631],
-//     [3.633, -0.107930757259, -0.700816114624],
-//     [3.6350000000000002, -0.120040425943, -0.711872912956],
-//     [3.672, -0.152306679493, -0.709795467823],
-//     [3.674, -0.158054757905, -0.73318936489],
-//     [3.6910000000000003, -0.149161831729, -0.719342630195],
-//     [3.698, -0.155450622153, -0.726648073246],
-//     [3.723, -0.117006288441, -0.745154502305],
-//     [3.741, -0.104548855448, -0.728401297692],
-//     [3.746, -0.0882034799869, -0.764763329267],
-//     [3.773, -0.138979298469, -0.804720654895],
-//     [3.7760000000000002, -0.151544046025, -0.792123766885],
-//     [3.803, -0.0940500275958, -0.783516323221],
-//     [3.821, -0.138709130902, -0.802963372606],
-//     [3.827, -0.152059451524, -0.755091520777],
-//     [3.855, -0.104376277191, -0.743065530452],
-//     [3.857, -0.0884520100182, -0.753384546279],
-//     [3.892, -0.123171341226, -0.756960670052],
-//     [3.898, -0.040082806437, -0.784347840209],
-//     [3.9170000000000003, -0.0876253173603, -0.819219639734],
-//     [3.935, -0.0430579824577, -0.826350599108],
-//     [3.94, -0.0779064468393, -0.826998706976],
-//     [3.98, -0.0270951708413, -0.812572381466],
-//     [3.983, -0.0271971095847, -0.78366083945],
-//     [3.991, 0.00767389574356, -0.780381853586],
-//     [4.019, -0.0142888258472, -0.803579496175],
-//     [4.021, -0.0397097774646, -0.810825740178],
-//     [4.049, 0.0145481445882, -0.822773527799],
-//     [4.0520000000000005, -0.0430078778181, -0.818051396632],
-//     [4.072, -0.00423406101655, -0.860191952725],
-//     [4.099, 0.00197882069917, -0.85965093922],
-//     [4.103, 0.0209627256494, -0.84163905729],
-//     [4.131, -0.0558021835039, -0.82058623879],
-//     [4.134, -0.0178256177774, -0.776055883246],
-//     [4.162, 0.00470082166988, -0.756183691505],
-//     [4.179, 0.00110048888346, -0.743224743067],
-//     [4.184, -0.0150582998325, -0.729104535373],
-//     [4.212, -0.0117828238451, -0.729287725082],
-//     [4.216, 0.0166983648175, -0.71556438532],
-//     [4.243, -0.03421464732, -0.721839159245],
-//     [4.246, -0.0056190281237, -0.713731822813],
-//     [4.2860000000000005, 0.00403100648779, -0.713970999173],
-//     [4.287, -0.00240263511856, -0.700216712533],
-//     [4.304, -0.0281956354581, -0.678353960359],
-//     [4.322, -0.0282536602977, -0.636433905592],
-//     [4.339, 0.0189125393385, -0.578110248217],
-//     [4.347, 0.0313313971861, -0.513818864518],
-//     [4.374, 0.0502621333422, -0.491379713722],
-//     [4.377, 0.0180435027936, -0.453982572689],
-//     [4.404, 0.0433064182716, -0.42381851778],
-//     [4.422, 0.0402299230717, -0.425922430492],
-//     [4.428, 0.0338768290451, -0.432184914243],
-//     [4.456, 0.0340046604125, -0.459762368834],
-//     [4.458, 0.0466806491906, -0.455860683906],
-//     [4.486, 0.0340687992931, -0.456591818335],
-//     [4.489, 0.0372274207573, -0.469441506817],
-//     [4.517, 0.0435309087568, -0.480000129224],
-//     [4.534, 0.0468793285706, -0.508559831564],
-//     [4.541, 0.0278223347948, -0.526537962187],
-//     [4.5680000000000005, 0.0503043015424, -0.542453419293],
-//     [4.571, 0.037833361834, -0.558028459154],
-//     [4.598, 0.0314268253557, -0.55424187353],
-//     [4.616, 0.0409367361761, -0.552468228961],
-//     [4.621, 0.00647186998196, -0.598315006261],
-//     [4.648, 0.0220324169141, -0.566085671407],
-//     [4.652, 0.00947199397323, -0.567175675832],
-//     [4.679, -0.0228294881391, -0.553355565364],
-//     [4.698, -0.00976341832289, -0.58247820174],
-//     [4.703, 0.0156622694124, -0.561751920485],
-//     [4.731, 0.0127420614304, -0.593208329748],
-//     [4.7330000000000005, 0.0314547225655, -0.556019080339],
-//     [4.761, 0.00575446129702, -0.514406916362],
-//     [4.764, 0.0148725742693, -0.470885167295],
-//     [4.793, 0.0468907741104, -0.451821475147],
-//     [4.811, 0.0243077416302, -0.431620678176],
-//     [4.814, 0.0335514740639, -0.402675987453],
-//     [4.8420000000000005, -0.0303327079026, -0.394681632538],
-//     [4.844, -0.00146620075095, -0.400637776311],
-//     [4.873, 0.00192226264464, -0.42092569463],
-//     [4.891, -0.0239164101969, -0.413551709744],
-//     [4.901, -0.0142052182961, -0.40855831571],
-//     [4.918, 0.0177715665593, -0.404608672361],
-//     [4.936, -0.0140807283322, -0.404800664286],
-//     [4.953, -0.0140888747731, -0.423107480909],
-//     [4.971, -0.0302421860241, -0.391056546969],
-//     [4.977, -0.0623376967428, -0.383795699293],
-//     [5.0040000000000004, -0.0690546514872, -0.361216141422],
-//     [5.007, -0.0717362674024, -0.400160956474],
-//     [5.034, -0.0814063761134, -0.401524630535],
-//     [5.0520000000000005, -0.103846657376, -0.408141070908],
-//     [5.058, -0.0909798793584, -0.427346604428],
-//     [5.086, -0.148009094506, -0.425633229856],
-//     [5.089, -0.150920603927, -0.484406460763],
-//     [5.117, -0.160375313375, -0.50553597937],
-//     [5.119, -0.147342000671, -0.535226209302],
-//     [5.147, -0.182405111147, -0.537643434904],
-//     [5.172, -0.176092878829, -0.545740396009],
-//     [5.173, -0.169856253276, -0.538191124631],
-//     [5.191, -0.221587157789, -0.48580832226],
-//     [5.218, -0.163949892792, -0.459226300705],
-//     [5.221, -0.211968122617, -0.472373509896],
-//     [5.248, -0.23701835818, -0.474992378055],
-//     [5.252, -0.179975473832, -0.453120789863],
-//     [5.28, -0.199506753605, -0.44231462887],
-//     [5.282, -0.176957512597, -0.45507014984],
-//     [5.299, -0.179899200371, -0.44481702705],
-//     [5.303, -0.154493197103, -0.457269561074],
-//     [5.331, -0.13515963206, -0.476380181427],
-//     [5.333, -0.173122341424, -0.513561005816],
-//     [5.361, -0.18855697171, -0.556319746114],
-//     [5.365, -0.160014294405, -0.55100003422],
-//     [5.392, -0.133765032743, -0.571340222337],
-//     [5.4110000000000005, -0.14356520167, -0.55538547189],
-//     [5.415, -0.141009488738, -0.528412500333],
-//     [5.442, -0.122428888092, -0.465239806058],
-//     [5.444, -0.116300474249, -0.443014325749],
-//     [5.472, -0.100157324036, -0.471349880972],
-//     [5.489, -0.100173384024, -0.437035178083],
-//     [5.503, -0.0839455180715, -0.442405185075],
-//     [5.516, -0.0806772223362, -0.461711917388],
-//     [5.527, -0.0685233362167, -0.421972449774],
-//     [5.546, -0.0778087723393, -0.418745254577],
-//     [5.573, -0.08397272524, -0.430056729108],
-//     [5.577, -0.0846531355687, -0.396743037882],
-//     [5.605, -0.0494281732803, -0.383849374713],
-//     [5.607, -0.0528809827477, -0.347652779933],
-//     [5.635, -0.059109605075, -0.343167359308],
-//     [5.6530000000000005, -0.0623137434655, -0.35713442372],
-//     [5.658, -0.0719517555578, -0.370236236697],
-//     [5.686, -0.0846162372831, -0.381395885953],
-//     [5.688, -0.0651678667828, -0.412209851792],
-//     [5.716, -0.122250824725, -0.464476685732],
-//     [5.719, -0.106329955766, -0.470093803887],
-//     [5.737, -0.0838214200402, -0.483942194175],
-//     [5.739, -0.0834491175137, -0.514190777549],
-//     [5.767, -0.112115498047, -0.549890169539],
-//     [5.769, -0.118660726411, -0.534931654335],
-//     [5.797, -0.0927906671882, -0.541168058819],
-//     [5.799, -0.095991539461, -0.539362984471],
-//     [5.817, -0.147196721293, -0.570462270262],
-//     [5.821, -0.0929144472053, -0.552449722136],
-//     [5.848, -0.146853780264, -0.569376998611],
-//     [5.851, -0.108449746818, -0.569317910561],
-//     [5.878, -0.120847155813, -0.594719402773],
-//     [5.882000000000001, -0.143403124399, -0.616612561476],
-//     [5.91, -0.0951105432719, -0.636402085659],
-//     [5.9270000000000005, -0.120466438837, -0.671070006415],
-//     [5.933, -0.0916860998811, -0.674240201907],
-//     [5.961, -0.139474341028, -0.706526693793],
-//     [5.963, -0.1201486803, -0.717232636207],
-//     [5.9910000000000005, -0.0916440851354, -0.712881404073],
-//     [6.008, -0.0692393030367, -0.713583033103],
-//     [6.015000000000001, -0.0275737828243, -0.716177958693],
-//     [6.042, -0.0724999020603, -0.71647771361],
-//     [6.046, -0.0534723671515, -0.681877530844],
-//     [6.073, -0.0696396226655, -0.664935130417],
-//     [6.0760000000000005, -0.0566105496575, -0.67814714066],
-//     [6.104, -0.0310645395687, -0.675323628584],
-//     [6.121, -0.0536181572493, -0.659012507302],
-//     [6.127, -0.0346566621817, -0.643132193732],
-//     [6.155, -0.0315635031307, -0.636048566841],
-//     [6.158, 0.0162247303541, -0.603936859072],
-//     [6.186, 0.0058167648683, -0.525739366813],
-//     [6.203, 0.0186276386358, -0.507235185354],
-//     [6.208, 0.00914673271253, -0.505745699801],
-//     [6.236, -0.0170579727243, -0.459778972226],
-//     [6.239, -0.00726326568876, -0.472352464577],
-//     [6.267, -0.0136450023562, -0.442604722947],
-//     [6.2700000000000005, -0.0428305959254, -0.433246440349],
-//     [6.299, -0.0269630776179, -0.386009438288],
-//     [6.316, -0.0270945045848, -0.401635470258],
-//     [6.321, -0.0237618439014, -0.402399367155],
-//     [6.348, -0.0555212992428, -0.406238974251],
-//     [6.351, -0.0175322021468, -0.394589881356],
-//     [6.3790000000000004, -0.0304569417026, -0.375817653556],
-//     [6.396, 0.00464802094311, -0.363523190205],
-//     [6.402, -0.0051013069695, -0.354094042686],
-//     [6.43, -0.0430044283314, -0.380098170204],
-//     [6.452, -0.0239768736203, -0.401964801334],
-//     [6.48, -0.0809186272653, -0.454913394445],
-//     [6.4830000000000005, -0.0806153353648, -0.47677695814],
-//     [6.511, -0.0712238754421, -0.460688607666],
-//     [6.513, -0.0931132361238, -0.550646281605],
-//     [6.541, -0.0897211591128, -0.562705163123],
-//     [6.558, -0.0958028166983, -0.597257632925],
-//     [6.565, -0.0893889360159, -0.608405026367],
-//     [6.593, -0.0729806896631, -0.656212246199],
-//     [6.595, -0.094777254084, -0.717510486627],
-//     [6.622, -0.0689185430921, -0.758155367223],
-//     [6.640000000000001, -0.0684024038139, -0.790640484403],
-//     [6.646, -0.0584993412826, -0.83681347024],
-//     [6.673, -0.0995896368989, -0.881464609756],
-//     [6.676, -0.0902712163421, -0.885421133262],
-//     [6.705, -0.102723354492, -0.927502052371],
-//     [6.722, -0.0578676778082, -0.963712860463],
-//     [6.727, -0.0637442215739, -0.98339934042],
-//     [6.747, -0.0668201196928, -1.00430184901],
-//     [6.765000000000001, -0.0470373021808, -1.04464762246],
-//     [6.782, -0.0376745286617, -1.04797866888],
-//     [6.788, -0.0312829114541, -1.05144513015],
-//     [6.806, -0.0407613989253, -1.0566003801],
-//     [6.808, -0.063260994613, -1.03497950689],
-//     [6.836, -0.0281006912562, -1.04001974668],
-//     [6.84, -0.0570371468448, -1.05953248231],
-//     [6.868, -0.0410560361011, -1.0402777492],
-//     [6.87, -0.0824431632671, -1.04212485954],
-//     [6.898000000000001, -0.0857903669748, -1.01118889809],
-//     [6.916, -0.0828666870524, -0.995637739007],
-//     [6.92, -0.0540375109217, -0.976556165048],
-//     [6.961, -0.047804814566, -0.980187613496],
-//     [6.963, -0.0415313958371, -0.961958900049],
-//     [6.9670000000000005, -0.0544000425386, -0.953776566409],
-//     [6.97, -0.0193479941878, -0.957384283013],
-//     [6.998, -0.0581991503351, -0.917998265742],
-//     [7.0, -0.0132027781081, -0.912866526502],
-//     [7.027, -0.0199349719057, -0.881109960045],
-//     [7.045, 0.021147545843, -0.817826311249],
-//     [7.051, 0.0340730978653, -0.83470851745],
-//     [7.078, 0.00824097787415, -0.807190341743],
-//     [7.081, 0.00476760435306, -0.786300940464],
-//     [7.1080000000000005, -0.0178011731812, -0.732866405059],
-//     [7.126, 0.0173102333622, -0.718584424884],
-//     [7.132000000000001, 0.017116281781, -0.727848415745],
-//     [7.16, -0.0118895292446, -0.701754063193],
-//     [7.162, -0.0208398998768, -0.765782156058],
-//     [7.19, -0.0432804863468, -0.765057389392],
-//     [7.1930000000000005, -0.0272706191532, -0.777190977017],
-//     [7.221, -0.071708867806, -0.77960978977],
-//     [7.238, -0.046510478099, -0.748458446626],
-//     [7.245, -0.0655716223597, -0.778292986064],
-//     [7.272, -0.0977357972805, -0.770294081364],
-//     [7.276, -0.10078608329, -0.81690693607],
-//     [7.305, -0.100972212987, -0.805546948738],
-//     [7.306, -0.106551449241, -0.872815967012],
-//     [7.333, -0.0776188447257, -0.896752565528],
-//     [7.352, -0.102512110684, -0.956223775245],
-//     [7.359, -0.0990987589434, -1.00038227393],
-//     [7.386, -0.121189093991, -1.0366695865],
-//     [7.4030000000000005, -0.0981420155761, -1.11228684296],
-//     [7.409, -0.065785715656, -1.12989673137],
-//     [7.4270000000000005, -0.110289269642, -1.16472554279],
-//     [7.445, -0.0877263783602, -1.18348450779],
-//     [7.462, -0.0497966736533, -1.16429101125],
-//     [7.48, -0.100659018012, -1.17024123677],
-//     [7.497, -0.122939273585, -1.19710173241],
-//     [7.515000000000001, -0.0810877014142, -1.20846955221],
-//     [7.518, -0.0586152319009, -1.21329107],
-//     [7.548, -0.0777032433911, -1.23589786994],
-//     [7.566, -0.0522544020688, -1.2276391208],
-//     [7.5680000000000005, -0.0424587494589, -1.25505491046],
-//     [7.596, -0.0649778879303, -1.24347360214],
-//     [7.598, -0.087058342687, -1.24517402997],
-//     [7.627, -0.0492157945569, -1.23006378191],
-//     [7.6450000000000005, -0.068821997306, -1.17216626684],
-//     [7.651, 0.014153196284, -1.14332874882],
-//     [7.68, -0.0118365506082, -1.10024567504],
-//     [7.681, 0.0326141971266, -1.06114916581],
-//     [7.708, 0.019332866336, -1.00219424743],
-//     [7.711, 0.0189668603007, -0.947346110043],
-//     [7.732, 0.0091337578371, -0.946392478898],
-//     [7.76, -0.000324312003583, -0.947534351315],
-//     [7.7620000000000005, 0.0660196908705, -0.908270010262],
-//     [7.79, 0.0406823807696, -0.886186942235],
-//     [7.793, 0.0853578173469, -0.877168683805],
-//     [7.821, 0.0689208858547, -0.80885542812],
-//     [7.838, 0.0847275679769, -0.773788098834],
-//     [7.845, 0.0851009490201, -0.798278469861],
-//     [7.872, 0.068680847921, -0.785646555031],
-//     [7.875, 0.0495197462526, -0.785353181485],
-//     [7.902, 0.059247877962, -0.819642932965],
-//     [7.92, 0.0146311048936, -0.845921337923],
-//     [7.926, 0.0436253914033, -0.876340496863],
-//     [7.956, 0.0404110184751, -0.877803908898],
-//     [7.961, 0.00894727899933, -0.908709946778],
-//     [7.981, 0.0313710267814, -0.92267806098],
-//     [7.987, -0.00328099941521, -0.98588352372],
-//     [8.015, -0.0222328342828, -1.0437281355],
-//     [8.032, -0.00956938725116, -1.0176348661],
-//     [8.037, -0.0416805584477, -0.995534614722],
-//     [8.065, -0.0132014778227, -0.946929255579],
-//     [8.067, -0.0200319461224, -0.889392471275],
-//     [8.095, -0.0136994690477, -0.867584877078],
-//     [8.112, -0.0427110078774, -0.8551906332],
-//     [8.118, 0.0177477128883, -0.843532069994],
-//     [8.146, -0.052617734342, -0.841758670983],
-//     [8.148, -0.0460682409316, -0.835558041399],
-//     [8.176, -0.0463063003342, -0.814038170098],
-//     [8.193, 0.00178333662561, -0.835340552972],
-//     [8.2, -0.014333718947, -0.831957014889],
-//     [8.227, -0.0495737129253, -0.803954310365],
-//     [8.23, -0.00847221587607, -0.760784947464],
-//     [8.257, 0.0232399860983, -0.72691584922],
-//     [8.275, 0.00402178794668, -0.703345875911],
-//     [8.281, 0.0295160715558, -0.712061353796],
-//     [8.308, -0.00227448130849, -0.721713652375],
-//     [8.311, 0.00742020456475, -0.71425749903],
-//     [8.338000000000001, -0.0149479221868, -0.742370382908],
-//     [8.356, 0.00480411551894, -0.807396610369],
-//     [8.362, -0.0333747588937, -0.84454088506],
-//     [8.39, -0.026575561698, -0.874264584705],
-//     [8.392, -0.0519378073036, -0.890139636628],
-//     [8.42, -0.0327597960033, -0.890472788547],
-//     [8.437, -0.0518977858432, -0.897414387916],
-//     [8.442, -0.0167413943532, -0.884447553433],
-//     [8.47, -0.0201557500236, -0.863512658366],
-//     [8.472, 0.0182580385189, -0.852313882611],
-//     [8.5, -0.0365107176858, -0.823916682063],
-//     [8.517, -0.0174601891742, -0.811868854892],
-//     [8.524000000000001, -0.0491687405206, -0.829361649811],
-//     [8.551, -0.0460144045056, -0.808784267154],
-//     [8.553, -0.0458858720345, -0.841036341981],
-//     [8.581, -0.0742934629728, -0.854897773502],
-//     [8.601, -0.0487577607963, -0.881862923099],
-//     [8.603, -0.0905299467054, -0.854941133627],
-//     [8.631, -0.0810986047205, -0.82511539824],
-//     [8.648, -0.0915187098551, -0.771022316865],
-//     [8.655, -0.05957272606, -0.76047664187],
-//     [8.682, -0.082870943992, -0.648691967526],
-//     [8.685, -0.0731319939697, -0.650735697287],
-//     [8.712, -0.102136058863, -0.628263089145],
-//     [8.73, -0.089236513549, -0.615374086808],
-//     [8.736, -0.0891857647539, -0.581258549475],
-//     [8.763, -0.118372559787, -0.552384446092],
-//     [8.766, -0.11836986021, -0.537497457018],
-//     [8.793000000000001, -0.118846606641, -0.489603976097],
-//     [8.796, -0.122058658127, -0.465133257641],
-//     [8.813, -0.138303198521, -0.443905349313],
-//     [8.816, -0.115673908281, -0.444455265424],
-//     [8.845, -0.11299574716, -0.402165556183],
-//     [8.846, -0.103908629352, -0.351533374447],
-//     [8.873, -0.151572705109, -0.383302046016],
-//     [8.891, -0.119961821707, -0.364876100138],
-//     [8.897, -0.13287144975, -0.349474728019],
-//     [8.925, -0.113942203862, -0.377090499561],
-//     [8.927, -0.132944180861, -0.380579726206],
-//     [8.955, -0.116486611967, -0.420309161896],
-//     [8.972, -0.173629841063, -0.46083611223],
-//     [8.978, -0.138241330871, -0.485415434655],
-//     [9.012, -0.185755935417, -0.52020812209],
-//     [9.013, -0.153428248459, -0.565000624223],
-//     [9.031, -0.175028049426, -0.604257782666],
-//     [9.049, -0.197347676903, -0.606373980303],
-//     [9.066, -0.181170471916, -0.657589998784],
-//     [9.083, -0.181449417431, -0.670162335861],
-//     [9.091, -0.18789631339, -0.654615904502],
-//     [9.118, -0.194290111246, -0.655027997141],
-//     [9.122, -0.178340513301, -0.656633140061],
-//     [9.15, -0.206961096937, -0.658343670306],
-//     [9.152000000000001, -0.21984799314, -0.651181697833],
-//     [9.17, -0.19790081558, -0.658472410059],
-//     [9.172, -0.159585427623, -0.639246666461],
-//     [9.200000000000001, -0.233273898897, -0.65300331756],
-//     [9.202, -0.197529892303, -0.68876959162],
-//     [9.23, -0.216151801064, -0.682315231173],
-//     [9.248, -0.165062227416, -0.705530971371],
-//     [9.253, -0.161962776295, -0.691995751422],
-//     [9.281, -0.203702268807, -0.667897314682],
-//     [9.283, -0.168768187588, -0.644416490034],
-//     [9.312, -0.165657373403, -0.59950636623],
-//     [9.315, -0.201040228806, -0.57971156088],
-//     [9.332, -0.10869157642, -0.512965921773],
-//     [9.335, -0.122871057487, -0.362443186132],
-//     [9.362, -0.0466719048705, -0.272128165946],
-//     [9.366, -0.121011107693, -0.176145194682],
-//     [9.394, -0.0955325549948, -0.134116121789],
-//     [9.396, -0.0575936042511, -0.108258109798],
-//     [9.423, -0.0769011857593, -0.0891377223127],
-//     [9.441, -0.0200687221391, -0.0443727067363],
-//     [9.447000000000001, -0.048778826176, -0.0536716544609],
-//     [9.475, -0.0170378238954, -0.0349741775202],
-//     [9.477, -0.0109598015574, -0.00856628170435],
-//     [9.505, -0.0647392304863, -0.0518253061684],
-//     [9.525, -0.067617835832, -0.083519372034],
-//     [9.527000000000001, -0.0828918914717, -0.156609948584],
-//     [9.555, -0.127008376063, -0.220463693233],
-//     [9.567, -0.183680481107, -0.299430772823],
-//     [9.58, -0.183093827348, -0.363236356692],
-//     [9.6, -0.202002123669, -0.428057787552],
-//     [9.617, -0.207906374432, -0.479873674555],
-//     [9.636000000000001, -0.224163563833, -0.478644373858],
-//     [9.653, -0.185189891911, -0.554748943853],
-//     [9.658, -0.246467754557, -0.546085060872],
-//     [9.686, -0.220765035946, -0.557127278669],
-//     [9.688, -0.236407089877, -0.573821908138],
-//     [9.716000000000001, -0.217461343018, -0.535756772306],
-//     [9.72, -0.230564461065, -0.579292844937],
-//     [9.747, -0.173016535143, -0.567526555005],
-//     [9.765, -0.265809568608, -0.557908830403],
-//     [9.771, -0.258661844418, -0.575170679011],
-//     [9.799, -0.236157217441, -0.580618647337],
-//     [9.801, -0.27736739259, -0.578162165824],
-//     [9.831, -0.258064725231, -0.585376740614],
-//     [9.849, -0.24887978223, -0.611554211772],
-//     [9.852, -0.27780539472, -0.578689542639],
-//     [9.88, -0.274705415445, -0.53957903256],
-//     [9.882, -0.220294415329, -0.556064989083],
-//     [9.91, -0.227061953314, -0.546171538589],
-//     [9.927, -0.208140705874, -0.530542762481],
-//     [9.933, -0.22085391035, -0.534652299163],
-//     [9.961, -0.271908453157, -0.530560464933],
-//     [9.965, -0.214613340444, -0.511998189164],
-//     [9.993, -0.214732889584, -0.477180610304],
-//     [9.995000000000001, -0.256723594463, -0.518805007294],
-//     [10.022, -0.22716323907, -0.530146696693],
-//     [10.040000000000001, -0.264755236559, -0.562925342138],
-//     [10.046, -0.210958186206, -0.54691757453],
-//     [10.074, -0.213660588976, -0.582687908829],
-//     [10.077, -0.178446695739, -0.6233024301],
-//     [10.105, -0.196774711811, -0.64700635675],
-//     [10.122, -0.187570440726, -0.661985978256],
-//     [10.129, -0.187534452956, -0.697049485304],
-//     [10.156, -0.20299788821, -0.738677914135],
-//     [10.159, -0.186436416181, -0.770207148891],
-//     [10.186, -0.186601015963, -0.805608001132],
-//     [10.19, -0.192714352268, -0.831219777424],
-//     [10.207, -0.202298798163, -0.825140865094],
-//     [10.21, -0.157025312354, -0.842850333025],
-//     [10.237, -0.180075929578, -0.817220951722],
-//     [10.241, -0.167392657393, -0.842026699836],
-//     [10.269, -0.138947147659, -0.848671037667],
-//     [10.271, -0.113645114388, -0.82457286019],
-//     [10.292, -0.10063725447, -0.846319588729],
-//     [10.32, -0.0680631984543, -0.874046510004],
-//     [10.321, -0.128597789021, -0.86022486224],
-//     [10.349, -0.128104128609, -0.897916428939],
-//     [10.351, -0.140831927972, -0.904261659218],
-//     [10.379, -0.138017679278, -0.862235831834],
-//     [10.396, -0.103240536427, -0.843603240778],
-//     [10.402000000000001, -0.106365522913, -0.872845114091],
-//     [10.431000000000001, -0.122303419629, -0.866732177371],
-//     [10.435, -0.0837752381448, -0.890907662009],
-//     [10.464, -0.0644702095801, -0.90816235068],
-//     [10.481, -0.0899173070873, -0.935676755042],
-//     [10.483, -0.0863959814084, -0.957133079184],
-//     [10.511000000000001, -0.0736580738838, -0.961830890653],
-//     [10.513, -0.0899393305673, -0.931312857568],
-//     [10.541, -0.0767681293139, -0.965675451657],
-//     [10.559000000000001, -0.0736098010187, -0.952616135031],
-//     [10.565, -0.0287053578345, -0.977076394262],
-//     [10.592, -0.0383739476462, -0.971532217817],
-//     [10.595, 0.0196441397884, -1.03239895836],
-//     [10.622, -0.0349700951717, -0.986548138575],
-//     [10.64, 0.0255255175595, -0.996885085997],
-//     [10.646, 0.00941778829534, -0.970222811871],
-//     [10.674, -0.000312269152957, -0.933061244056],
-//     [10.676, 0.0347262145275, -0.924739338443],
-//     [10.704, -0.000301134551407, -0.923764895133],
-//     [10.721, -0.051378537724, -0.931748500907],
-//     [10.729000000000001, -0.0256511091936, -0.959162029497],
-//     [10.75, -0.00927304474161, -0.991842201175],
-//     [10.757, -0.0381256576878, -0.979108414942],
-//     [10.786, -0.0605060189737, -0.997076574264],
-//     [10.804, -0.0252451380509, -0.99292183328],
-//     [10.807, -0.0574090978601, -0.994411823214],
-//     [10.835, -0.0060646733249, -0.973188956974],
-//     [10.837, -0.0189736785839, -0.954263537231],
-//     [10.865, -0.0223865914377, -0.944053708548],
-//     [10.884, -0.0707697770954, -0.895720812656],
-//     [10.888, -0.0357999823154, -0.888461705069],
-//     [10.916, -0.0355344139141, -0.92887386608],
-//     [10.92, -0.0321591187022, -0.953866020972],
-//     [10.947000000000001, -0.0768422980317, -0.945311608067],
-//     [10.965, -0.0573249851154, -0.943075525472],
-//     [10.97, -0.0479904373895, -0.955193271831],
-//     [10.999, -0.0641839631141, -0.917227874074],
-//     [11.0, -0.0932680285031, -0.871282052134],
-//     [11.029, -0.0775275917628, -0.852048843406],
-//     [11.05, -0.106986951261, -0.799816922175],
-//     [11.051, -0.113840050187, -0.767143788991],
-//     [11.08, -0.0978698606686, -0.760392886505],
-//     [11.081, -0.113499760133, -0.799271644952],
-//     [11.109, -0.110101297917, -0.782691485994],
-//     [11.126, -0.112598548384, -0.801599290404],
-//     [11.132, -0.180292344291, -0.77420363013],
-//     [11.16, -0.169836750535, -0.811683999076],
-//     [11.163, -0.148025134859, -0.797708155943],
-//     [11.19, -0.0973011030719, -0.766930061746],
-//     [11.207, -0.209209433995, -0.742935012395],
-//     [11.214, -0.174754799594, -0.713344782532],
-//     [11.241, -0.178288366362, -0.673708814284],
-//     [11.244, -0.165522939827, -0.610625442917],
-//     [11.271, -0.175612778058, -0.557314517631],
-//     [11.275, -0.172553369255, -0.536095511747],
-//     [11.302, -0.17587942311, -0.511467263265],
-//     [11.32, -0.211387231585, -0.518470373655],
-//     [11.326, -0.188990477046, -0.48581901929],
-//     [11.354000000000001, -0.211893608501, -0.492447222242],
-//     [11.356, -0.198312473142, -0.509766614245],
-//     [11.384, -0.198004160804, -0.505003724253],
-//     [11.401, -0.226895277018, -0.491629366672],
-//     [11.407, -0.191803318615, -0.512181154495],
-//     [11.435, -0.20441954711, -0.550727214373],
-//     [11.437, -0.220688764998, -0.569228997731],
-//     [11.465, -0.24526849239, -0.612584208394],
-//     [11.482000000000001, -0.261274796491, -0.641324001467],
-//     [11.489, -0.247842621195, -0.69355598541],
-//     [11.516, -0.270515592031, -0.727503424352],
-//     [11.519, -0.231304338204, -0.782602618732],
-//     [11.546, -0.237850645337, -0.795419185527],
-//     [11.572000000000001, -0.266707557102, -0.789271839624],
-//     [11.574, -0.234992753574, -0.774990867966],
-//     [11.591000000000001, -0.25367619059, -0.796894777853],
-//     [11.609, -0.241668295404, -0.733381796285],
-//     [11.626, -0.190286903728, -0.751468377216],
-//     [11.647, -0.215823739321, -0.747381087773],
-//     [11.654, -0.177324065195, -0.750559897257],
-//     [11.672, -0.222154122766, -0.754668142075],
-//     [11.69, -0.222099005749, -0.731574272719],
-//     [11.707, -0.202588704819, -0.754049515014],
-//     [11.711, -0.180092226916, -0.766457561578],
-//     [11.729000000000001, -0.195852426223, -0.763690670862],
-//     [11.731, -0.189655718727, -0.760169239011],
-//     [11.759, -0.170070545069, -0.782610157511],
-//     [11.762, -0.199619848144, -0.74898293685],
-//     [11.790000000000001, -0.167882177357, -0.73349142146],
-//     [11.808, -0.104810432509, -0.663694912261],
-//     [11.814, -0.0766542620736, -0.601865822206],
-//     [11.841000000000001, -0.0929599179885, -0.590191033287],
-//     [11.845, -0.0768555405043, -0.563861487395],
-//     [11.865, -0.0670406632789, -0.595162316585],
-//     [11.882, -0.120773465622, -0.637338183576],
-//     [11.9, -0.0917593901204, -0.661418578495],
-//     [11.917, -0.0758535687268, -0.631698488627],
-//     [11.961, -0.057208935457, -0.576093014772],
-//     [11.97, -0.0698143673987, -0.592653050112],
-//     [11.971, -0.0796310658703, -0.571766699338],
-//     [11.985, -0.0833353166539, -0.55645847414],
-//     [12.002, -0.124714847079, -0.561387731702],
-//     [12.007, -0.054097879885, -0.586345559733],
-//     [12.035, -0.105016319264, -0.636608564768],
-//     [12.037, -0.108055894875, -0.637990042666],
-//     [12.065, -0.130768237323, -0.617844225144],
-//     [12.082, -0.117967845105, -0.645352894493],
-//     [12.089, -0.159671749895, -0.623299590622],
-//     [12.116, -0.0924199602572, -0.637858890152],
-//     [12.119, -0.0952824374114, -0.653108662232],
-//     [12.147, -0.121013033376, -0.634362513928],
-//     [12.149000000000001, -0.133970103307, -0.626141323759],
-//     [12.176, -0.143359694773, -0.632517367646],
-//     [12.194, -0.104649650695, -0.660934378836],
-//     [12.201, -0.114535459085, -0.656809697518],
-//     [12.229000000000001, -0.130647056305, -0.656245012722],
-//     [12.23, -0.111148139686, -0.673487590087],
-//     [12.258000000000001, -0.114318955319, -0.666425966634],
-//     [12.277000000000001, -0.0796310658703, -0.624528187237],
-//     [12.281, -0.0568997694, -0.656368182922],
-//     [12.309000000000001, -0.056982815739, -0.664954039636],
-//     [12.311, -0.0822924064722, -0.681101612546],
-//     [12.339, -0.0824337952291, -0.672224918715],
-//     [12.356, -0.053488358818, -0.679188300565],
-//     [12.362, -0.0979654702024, -0.704867669444],
-//     [12.39, -0.0664734895651, -0.639902845736],
-//     [12.392, -0.040990021888, -0.629668786667],
-//     [12.42, -0.0221143329961, -0.59474549496],
-//     [12.437, -0.0448473173882, -0.544000220594],
-//     [12.444, -0.0193251912967, -0.547663931619],
-//     [12.471, -0.0226421100767, -0.553024078499],
-//     [12.474, -0.00973566641409, -0.54842692873],
-//     [12.501, -0.022452586828, -0.551622948595],
-//     [12.519, -0.041788379627, -0.551892994636],
-//     [12.525, -0.0477039912817, -0.586745463817],
-//     [12.552, -0.0992386752485, -0.568063753403],
-//     [12.555, -0.0382253074571, -0.586105137841],
-//     [12.582, -0.0288733943427, -0.582222204635],
-//     [12.6, -0.0705399612825, -0.580494549476],
-//     [12.606, -0.0543354040057, -0.604297130114],
-//     [12.634, -0.0986728799152, -0.617352674576],
-//     [12.636000000000001, -0.127432437825, -0.619456469557],
-//     [12.664, -0.140191197419, -0.599374895118],
-//     [12.681000000000001, -0.115154377706, -0.551151771536],
-//     [12.687, -0.153604534539, -0.528577830631],
-//     [12.715, -0.156641467016, -0.545835582252],
-//     [12.717, -0.137944026603, -0.527664898071],
-//     [12.745000000000001, -0.156430791732, -0.571706441898],
-//     [12.749, -0.137172818607, -0.582508892238],
-//     [12.766, -0.178328283275, -0.606614745636],
-//     [12.769, -0.206546052455, -0.641201312356],
-//     [12.797, -0.197305730201, -0.62979607279],
-//     [12.8, -0.177820667433, -0.640362213008],
-//     [12.827, -0.168915947342, -0.637264098505],
-//     [12.83, -0.223542498311, -0.666175329274],
-//     [12.852, -0.226927659674, -0.638098683274],
-//     [12.88, -0.197962525966, -0.65911693207],
-//     [12.882, -0.168842046087, -0.65594698445],
-//     [12.911, -0.235265871855, -0.694023262264],
-//     [12.914, -0.209589077783, -0.724746856609],
-//     [12.939, -0.234681719412, -0.792373882749],
-//     [12.956, -0.237038000177, -0.864486152346],
-//     [12.962, -0.252620433216, -0.919527590443],
-//     [12.982000000000001, -0.249054838369, -0.951154062526],
-//     [13.0, -0.226641616328, -0.960799410105],
-//     [13.017, -0.255747996763, -0.960460437832],
-//     [13.035, -0.230105133693, -0.933252834242],
-//     [13.052, -0.281490145698, -0.925216945489],
-//     [13.07, -0.220478783422, -0.923130467676],
-//     [13.074, -0.205569945235, -0.914652224688],
-//     [13.101, -0.195681438084, -0.87554446673],
-//     [13.119, -0.224206931437, -0.892606443087],
-//     [13.125, -0.157799239764, -0.870263869372],
-//     [13.152000000000001, -0.202142936784, -0.883621891538],
-//     [13.155000000000001, -0.176322898362, -0.874792974935],
-//     [13.186, -0.176931207021, -0.854279397616],
-//     [13.187, -0.160548671813, -0.882567923838],
-//     [13.216000000000001, -0.163652755132, -0.875652178368],
-//     [13.234, -0.141294897649, -0.872585009178],
-//     [13.236, -0.150489255126, -0.872419627813],
-//     [13.264000000000001, -0.146907761584, -0.892452975776],
-//     [13.267, -0.124288142388, -0.940544294154],
-//     [13.295, -0.124835895352, -0.920509903799],
-//     [13.312000000000001, -0.115583379068, -0.888594271649],
-//     [13.319, -0.109202920283, -0.901339452271],
-//     [13.346, -0.110019137383, -0.854851461051],
-//     [13.35, -0.119448877735, -0.851469033674],
-//     [13.378, -0.132002459016, -0.881146669193],
-//     [13.38, -0.125697604876, -0.869542813466],
-//     [13.407, -0.131496627247, -0.905086986191],
-//     [13.425, -0.102566691988, -0.907516115451],
-//     [13.431000000000001, -0.128289826826, -0.883678912365],
-//     [13.459, -0.0906915659465, -0.853299693632],
-//     [13.462, -0.0874041851938, -0.84644953205],
-//     [13.49, -0.112904621593, -0.843147131216],
-//     [13.501, -0.0617474068943, -0.838461743763],
-//     [13.521, -0.0906834820477, -0.836860183325],
-//     [13.539, -0.0525847162306, -0.805984435939],
-//     [13.542, -0.0905221164748, -0.831008129908],
-//     [13.57, -0.0873240294428, -0.836533927681],
-//     [13.589, -0.099874466143, -0.87471994561],
-//     [13.594, -0.0804977775493, -0.914193532248],
-//     [13.621, -0.121863347449, -0.935883792682],
-//     [13.625, -0.131129911365, -0.966640725195],
-//     [13.652000000000001, -0.0765943732161, -1.00565843831],
-//     [13.665000000000001, -0.0954511493187, -1.03267885392],
-//     [13.677, -0.0570989120505, -1.02838541402],
-//     [13.696, -0.105050761085, -1.05109391829],
-//     [13.714, -0.0503557557349, -1.07303212322],
-//     [13.731, -0.0663500899019, -1.09178547274],
-//     [13.749, -0.0751901276971, -1.12890260849],
-//     [13.755, -0.0526306129418, -1.1827232496],
-//     [13.782, -0.0655223988561, -1.16418825521],
-//     [13.786, -0.0518924163372, -1.24013939397],
-//     [13.814, -0.0740344557936, -1.24517888453],
-//     [13.831, -0.0677944345308, -1.23240758063],
-//     [13.837, -0.0163323125646, -1.24061595435],
-//     [13.865, -0.0131151748147, -1.25063744156],
-//     [13.869, -0.0582418784564, -1.21528149165],
-//     [13.896, -0.0231061584549, -1.21331174536],
-//     [13.899000000000001, -0.0165705105305, -1.21337239143],
-//     [13.926, -0.0291605342617, -1.22962327989],
-//     [13.944, -0.0418398759723, -1.24053886868],
-//     [13.950000000000001, -0.0292608870281, -1.25338550813],
-//     [13.977, -0.0326030734729, -1.25425121931],
-//     [13.98, -0.0417791111421, -1.28401773905],
-//     [14.007, -0.0639302496459, -1.30430338212],
-//     [14.025, -0.0889109330893, -1.32366664791],
-//     [14.031, -0.0663388328007, -1.3655871155],
-//     [14.059000000000001, -0.0410446512664, -1.32881352804],
-//     [14.062000000000001, -0.00315701566271, -1.31538824365],
-//     [14.09, -0.047894078843, -1.28865151023],
-//     [14.092, -0.0193679416405, -1.2652077302],
-//     [14.132, -0.0229416014702, -1.25115421089],
-//     [14.134, -0.0135351165301, -1.23393299392],
-//     [14.151, -0.0230170316909, -1.24427868561],
-//     [14.169, -0.0323892935281, -1.25758749985],
-//     [14.175, -0.0771041632556, -1.27223100322],
-//     [14.202, -0.0771041632556, -1.26846079412],
-//     [14.22, -0.0610008928871, -1.28434030751],
-//     [14.225, -0.0798538828209, -1.30969729383],
-//     [14.253, -0.0480754021466, -1.28918239365],
-//     [14.255, -0.0636318179956, -1.35458746526],
-//     [14.283, -0.11087133704, -1.38093245164],
-//     [14.3, -0.0819575665132, -1.38567337806],
-//     [14.306000000000001, -0.11441161592, -1.3548771162],
-//     [14.334, -0.0990817063834, -1.33177241981],
-//     [14.336, -0.108692747392, -1.32918857771],
-//     [14.364, -0.102717203011, -1.28800849077],
-//     [14.381, -0.0998898415008, -1.23834505197],
-//     [14.388, -0.116112987429, -1.21572788757],
-//     [14.415000000000001, -0.0775176723542, -1.22577674349],
-//     [14.424, -0.087058342687, -1.23385206282],
-//     [14.441, -0.0939042531017, -1.2046419263],
-//     [14.449, -0.0972052578488, -1.19309404892],
-//     [14.476, -0.0972539846683, -1.18702973566],
-//     [14.494, -0.119418061346, -1.19248967313],
-//     [14.5, -0.0808268317545, -1.21156308154],
-//     [14.529, -0.0904286567218, -1.21699857716],
-//     [14.530000000000001, -0.0935319258035, -1.25228430281],
-//     [14.558, -0.125107278727, -1.27829114145],
-//     [14.575000000000001, -0.0643444842079, -1.28238877001],
-//     [14.581, -0.089242305967, -1.33646617781],
-//     [14.61, -0.0757003911584, -1.37735824395],
-//     [14.611, -0.092006554945, -1.36926758905],
-//     [14.639000000000001, -0.0123935470285, -1.35692662293],
-//     [14.656, -0.0151826631321, -1.36916057421],
-//     [14.664, -0.0850274447016, -1.40893203588],
-//     [14.691, -0.0465896048768, -1.41576396305],
-//     [14.694, -0.0653049923731, -1.45313105677],
-//     [14.721, -0.0338300037162, -1.43423249532],
-//     [14.733, -0.0337508787104, -1.41235581683],
-//     [14.75, -0.0370745090418, -1.38682807702],
-//     [14.767, -0.0016083095451, -1.39996564357],
-//     [14.775, -0.00240761581875, -1.3454178593],
-//     [14.803, -0.00582315189484, -1.33473359984],
-//     [14.806000000000001, -0.0446207236351, -1.28500431082],
-//     [14.831, -0.00648105056203, -1.28169599385],
-//     [14.849, -0.0222010645422, -1.28538487929],
-//     [14.866, -0.0475471894324, -1.31089052917],
-//     [14.884, -0.034686023668, -1.30986908065],
-//     [14.887, -0.0921061474644, -1.31773757545],
-//     [14.915000000000001, -0.0700482910338, -1.28335997489],
-//     [14.934000000000001, -0.13096903276, -1.27304660871],
-//     [14.938, -0.111924570326, -1.27051107154],
-//     [14.965, -0.0770821675242, -1.22814436592],
-//     [14.969000000000001, -0.0452332757795, -1.22105571939],
-//     [14.996, -0.0612471599491, -1.22383137356],
-//     [15.014000000000001, -0.0868810596277, -1.22750782347],
-//     [15.019, -0.0675629262195, -1.25288070529],
-//     [15.048, -0.0998660076772, -1.20624590366],
-//     [15.049, -0.0774444926345, -1.19370536652],
-//     [15.076, -0.119083662876, -1.19372072922],
-//     [15.094000000000001, -0.0647973910697, -1.20352803883],
-//     [15.1, -0.0708850941187, -1.2130165821],
-//     [15.129, -0.10603259167, -1.22175378445],
-//     [15.131, -0.109749289261, -1.18822175622],
-//     [15.159, -0.0937923233386, -1.16323821462],
-//     [15.162, -0.0908482109531, -1.14031152734],
-//     [15.189, -0.120262554172, -1.10815745531],
-//     [15.206, -0.0884580733637, -1.10310691777],
-//     [15.213000000000001, -0.111049513102, -1.05836333083],
-//     [15.24, -0.114690747864, -1.03044963851],
-//     [15.243, -0.0989045409326, -0.993182555926],
-//     [15.27, -0.0705368023394, -0.945087797719],
-//     [15.288, -0.0900705386314, -0.898744154984],
-//     [15.295, -0.0803943752763, -0.899965841903],
-//     [15.323, -0.0677156761242, -0.89159330105],
-//     [15.325000000000001, -0.0802705709395, -0.897446782618],
-//     [15.354000000000001, -0.0866434685704, -0.88180963468],
-//     [15.356, -0.093041976669, -0.898819311276],
-//     [15.384, -0.0547897748131, -0.906178611958],
-//     [15.414, -0.0353563208991, -0.933252961255],
-//     [15.415000000000001, -0.0926489542827, -0.939194644016],
-//     [15.433, -0.111697281562, -0.969598598392],
-//     [15.439, -0.102049039603, -0.995222969092],
-//     [15.466000000000001, -0.111497543892, -1.01957922654],
-//     [15.484, -0.108284513143, -1.00654704206],
-//     [15.487, -0.079663323696, -1.00173057538],
-//     [15.515, -0.089392349075, -0.987695217592],
-//     [15.519, -0.10199455887, -1.00623485884],
-//     [15.546000000000001, -0.0764360283782, -0.984997023255],
-//     [15.564, -0.0864737094617, -0.967365024402],
-//     [15.57, -0.105794970951, -0.945235771841],
-//     [15.598, -0.0674603849011, -0.940139329147],
-//     [15.603, -0.0545376804467, -0.945693445266],
-//     [15.625, -0.0287564011888, -0.968462507499],
-//     [15.631, -0.0318420664858, -0.994798546918],
-//     [15.659, -0.0512192328061, -0.964545218656],
-//     [15.676, -0.0479752291437, -0.949447552648],
-//     [15.683, -0.0225399006484, -0.950099374541],
-//     [15.71, -0.0512288199883, -0.921347297392],
-//     [15.713000000000001, -0.0963805919524, -0.889217634103],
-//     [15.74, -0.0520620855618, -0.848600431302],
-//     [15.758000000000001, -0.036564653675, -0.78536314532],
-//     [15.763, -0.0271819355652, -0.774541791983],
-//     [15.790000000000001, -0.0650651626266, -0.798199798129],
-//     [15.808, -0.0366260371936, -0.789437080611],
-//     [15.814, -0.0398962935349, -0.76927859547],
-//     [15.841000000000001, -0.0594958427328, -0.714724086211],
-//     [15.845, -0.0373547395474, -0.673051323605],
-//     [15.873000000000001, -0.0346102817736, -0.632648930618],
-//     [15.874, -0.0379046279863, -0.627270502391],
-//     [15.901, -0.00294121272007, -0.568963152],
-//     [15.919, -0.0607390403442, -0.53427391875],
-//     [15.925, -0.051328874541, -0.521340419456],
-//     [15.953000000000001, -0.0960902036389, -0.51982552557],
-//     [15.97, -0.0929991862273, -0.539148820192],
-//     [15.975, -0.0542240230663, -0.533089401604],
-//     [16.003, -0.0830637855509, -0.567276139211],
-//     [16.006, -0.0571896544874, -0.596824736483],
-//     [16.034, -0.136764399007, -0.652277185872],
-//     [16.036, -0.180986695551, -0.712188380905],
-//     [16.064, -0.193060553167, -0.771397986732],
-//     [16.081, -0.148240445504, -0.79895420627],
-//     [16.088, -0.158188622676, -0.791345759535],
-//     [16.115000000000002, -0.173923277373, -0.822558893727],
-//     [16.118000000000002, -0.182609640274, -0.854857524403],
-//     [16.145, -0.19857052424, -0.83344097651],
-//     [16.163, -0.157332182489, -0.82208128062],
-//     [16.169, -0.208749689567, -0.857159382043],
-//     [16.196, -0.208882509003, -0.854190988297],
-//     [16.199, -0.195681438084, -0.879315133999],
-//     [16.226, -0.226706878422, -0.922545136207],
-//     [16.244, -0.229525893476, -0.973242340315],
-//     [16.249, -0.251977692879, -1.04969281778],
-//     [16.276, -0.248555298891, -1.05456010339],
-//     [16.279, -0.241569251482, -1.05643449446],
-//     [16.306, -0.19690704498, -1.09627651003],
-//     [16.324, -0.206568068761, -1.10241180006],
-//     [16.330000000000002, -0.228874490639, -1.09548439596],
-//     [16.358, -0.200353871709, -1.09897527622],
-//     [16.36, -0.187685309206, -1.08870238809],
-//     [16.388, -0.194259593429, -1.06532228398],
-//     [16.405, -0.152706367491, -1.07949936381],
-//     [16.411, -0.124331572927, -1.02933259464],
-//     [16.439, -0.10808154929, -1.0528360692],
-//     [16.441, -0.101460788387, -1.07210537491],
-//     [16.469, -0.152553448083, -1.0706262925],
-//     [16.486, -0.101707518109, -1.00533917647],
-//     [16.493000000000002, -0.130385816391, -1.01139035384],
-//     [16.52, -0.115000749427, -0.981299319666],
-//     [16.523, -0.124457467934, -0.979559550521],
-//     [16.55, -0.0897089941844, -0.962743126227],
-//     [16.568, -0.0990487202029, -0.975826801078],
-//     [16.575, -0.0576572401938, -0.96044914397],
-//     [16.603, -0.0927218881324, -0.963293882507],
-//     [16.605, -0.0640358476283, -0.957258679574],
-//     [16.633, -0.0829481365657, -0.979260392599],
-//     [16.635, -0.0384360192586, -0.945145637015],
-//     [16.663, -0.0485877633124, -0.882440288756],
-//     [16.68, -0.0647691340943, -0.879047740736],
-//     [16.695, -0.0743702575559, -0.874042440419],
-//     [16.713, -0.0296229448023, -0.86957079483],
-//     [16.716, -0.0518178123103, -0.896866217703],
-//     [16.744, -0.0388705107211, -0.922468235395],
-//     [16.761, -0.0707453370901, -0.901565721881],
-//     [16.768, -0.0329464301489, -0.879116662323],
-//     [16.795, -0.0681501451722, -0.865115010322],
-//     [16.798000000000002, -0.0553049420032, -0.865976698353],
-//     [16.825, -0.109362172393, -0.90273838374],
-//     [16.843, -0.077344764206, -0.906042674801],
-//     [16.849, -0.0933838595604, -0.918489746245],
-//     [16.876, -0.0864496573304, -0.956645499284],
-//     [16.88, -0.0894877511219, -0.975447556222],
-//     [16.908, -0.0995779360112, -0.948108720939],
-//     [16.925, -0.0481611182221, -0.94189156166],
-//     [16.93, -0.0422511294713, -0.879234371167],
-//     [16.958000000000002, -0.0394095546666, -0.841595401533],
-//     [16.961000000000002, -0.0522743940077, -0.828652561309],
-//     [17.0, -0.0205283488799, -0.815422214358],
-//     [17.001, -0.010724995696, -0.819994108473],
-//     [17.02, -0.0230398392577, -0.865033919232],
-//     [17.038, -0.0870055634218, -0.868334311259],
-//     [17.043, -0.0550153716824, -0.865750224663],
-//     [17.07, -0.0361556355101, -0.834508537224],
-//     [17.074, -0.0329257970438, -0.862399543265],
-//     [17.101, -0.0713693347853, -0.851762988413],
-//     [17.119, -0.0137241648395, -0.872170177168],
-//     [17.125, -0.0261996524991, -0.903931365555],
-//     [17.153, -0.0484175656343, -0.916935390288],
-//     [17.155, 0.0186319645459, -0.927841715763],
-//     [17.183, -0.0262973638374, -0.908531869846],
-//     [17.186, -0.0392851330789, -0.871828374631],
-//     [17.214, -0.0519515740606, -0.902872161557],
-//     [17.233, -0.0354351435689, -0.915515588836],
-//     [17.238, -0.0291299110102, -0.93062893329],
-//     [17.265, -0.0799268192374, -0.93834536094],
-//     [17.268, -0.0252951825594, -0.978551134978],
-//     [17.295, -0.0601865847232, -1.01692017567],
-//     [17.313, -0.0445608527974, -0.996071187253],
-//     [17.319, -0.0792091643356, -1.05367667686],
-//     [17.346, -0.0698693180069, -1.04699522738],
-//     [17.349, -0.0920912512752, -1.04868917344],
-//     [17.376, -0.0405665654761, -1.08746220409],
-//     [17.396, -0.0662534566863, -1.0637413244],
-//     [17.399, -0.0697728686072, -1.05163122824],
-//     [17.426000000000002, -0.11803055228, -1.01703859332],
-//     [17.429000000000002, -0.0796352831398, -1.01369262168],
-//     [17.456, -0.0702071297357, -0.994988795521],
-//     [17.474, -0.0445638633544, -0.993845559472],
-//     [17.48, -0.0570160246278, -1.01269390155],
-//     [17.508, -0.0185743866548, -1.05771128446],
-//     [17.51, -0.0502655537191, -1.08191570931],
-//     [17.538, -0.0535322710151, -1.08194233707],
-//     [17.555, -0.0562931395786, -1.09810304845],
-//     [17.561, -0.0816186554756, -1.12510088763],
-//     [17.589, -0.0530758244502, -1.10600219456],
-//     [17.591, -0.0589833353193, -1.13734112424],
-//     [17.619, -0.0560688424574, -1.13414543334],
-//     [17.636, -0.0880727630654, -1.10867360379],
-//     [17.643, -0.0719313487392, -1.12451835693],
-//     [17.67, -0.0787892624578, -1.1062130123],
-//     [17.673000000000002, -0.069099676325, -1.1221135709],
-//     [17.7, -0.0946647914751, -1.13715336785],
-//     [17.718, -0.13927756782, -1.13865432179],
-//     [17.73, -0.138755447715, -1.167872358],
-//     [17.744, -0.0810912405241, -1.1877202574],
-//     [17.755, -0.0934365709362, -1.20100662024],
-//     [17.783, -0.0968920189024, -1.18981999477],
-//     [17.785, -0.122254203613, -1.18976647402],
-//     [17.813, -0.13241958292, -1.1562053265],
-//     [17.830000000000002, -0.113434291992, -1.14383564392],
-//     [17.837, -0.113767109272, -1.12470692744],
-//     [17.864, -0.155335486044, -1.12933860604],
-//     [17.866, -0.129754880852, -1.14633063781],
-//     [17.894000000000002, -0.164834988478, -1.15928351023],
-//     [17.911, -0.129735621226, -1.15804038103],
-//     [17.918, -0.113569486082, -1.15046694564],
-//     [17.945, -0.180392499388, -1.10735220764],
-//     [17.948, -0.164474668901, -1.1234536327],
-//     [17.975, -0.170752553425, -1.10840788512],
-//     [17.993000000000002, -0.0976200863024, -1.09594794277],
-//     [17.999, -0.0984196280578, -1.07070034822],
-//     [18.026, -0.069827014001, -1.06378967028],
-//     [18.03, -0.0891504051631, -1.03505238438],
-//     [18.059, -0.0156834288527, -1.00538709563],
-//     [18.076, -0.04121928119, -1.01255313199],
-//     [18.080000000000002, -0.0410233917252, -1.02922214407],
-//     [18.109, -0.0377504263896, -1.02217717264],
-//     [18.11, -0.0473860466842, -1.02230715766],
-//     [18.138, -0.0280884929758, -1.04940774558],
-//     [18.141000000000002, -0.0122209059055, -1.0357661089],
-//     [18.169, -0.0311504059731, -1.07542325393],
-//     [18.186, -0.00555997275988, -1.06074590088],
-//     [18.191, 0.0236221055925, -1.08844077908],
-//     [18.219, 0.0205346537153, -1.11966395592],
-//     [18.223, 0.0497072428019, -1.17423357819],
-//     [18.25, 0.0434543743306, -1.18927604555],
-//     [18.268, 0.0498186219634, -1.18357418955],
-//     [18.274, 0.0722594973867, -1.16763377702],
-//     [18.301000000000002, 0.0691974379856, -1.14304472578],
-//     [18.305, 0.0941601172768, -1.10317723767],
-//     [18.333000000000002, 0.128963165199, -1.07508490387],
-//     [18.35, 0.090653435145, -1.07905574533],
-//     [18.355, 0.097009863998, -1.05234139467],
-//     [18.383, 0.118870087753, -1.03705267366],
-//     [18.385, 0.0587150677229, -1.05534614043],
-//     [18.413, 0.0875557724614, -1.03144359208],
-//     [18.433, 0.0619027117873, -1.07182886547],
-//     [18.435, 0.0521864811832, -1.06314728979],
-//     [18.463, 0.0592444289049, -1.11029296418],
-//     [18.465, 0.062499136597, -1.09760516664],
-//     [18.493000000000002, 0.0653101054438, -1.09345385955],
-//     [18.51, 0.0457131136342, -1.05426942158],
-//     [18.515, 0.0456542040565, -1.02946565967],
-//     [18.543, 0.0458190329701, -1.04306250962],
-//     [18.56, 0.0134737562344, -1.00759482451],
-//     [18.566, -0.00249353751295, -1.01117009504],
-//     [18.594, -0.0217958317255, -1.01925234026],
-//     [18.596, 0.00384311027314, -1.01809163387],
-//     [18.624, -0.01232176462, -1.00467573389],
-//     [18.642, -0.0504688953806, -1.02594494305],
-//     [18.648, -0.06652703483, -1.01000311701],
-//     [18.675, -0.0565405554929, -1.04732349444],
-//     [18.678, -0.0822002604173, -1.03753317032],
-//     [18.705000000000002, -0.0760515651443, -1.01673852411],
-//     [18.723, -0.0858899179833, -0.996719188287],
-//     [18.728, -0.159444608189, -0.981223671115],
-//     [18.755, -0.130756850404, -0.962322903819],
-//     [18.758, -0.14006360274, -0.993601057185],
-//     [18.785, -0.190835297471, -0.976880163927],
-//     [18.803, -0.140746844005, -0.944817776823],
-//     [18.809, -0.166396067935, -0.908487351742],
-//     [18.836000000000002, -0.172832198067, -0.889878042656],
-//     [18.84, -0.112594651172, -0.862076291647],
-//     [18.868000000000002, -0.135140926421, -0.820894352312],
-//     [18.886, -0.195906280124, -0.796840930597],
-//     [18.891000000000002, -0.151592946248, -0.790121933219],
-//     [18.918, -0.145588417711, -0.732966353257],
-//     [18.92, -0.168126297709, -0.712004894093],
-//     [18.948, -0.149261528721, -0.694276638323],
-//     [18.965, -0.187499488418, -0.670194014419],
-//     [18.971, -0.124275341953, -0.65333354554],
-//     [18.999, -0.133992349745, -0.617260380956],
-//     [19.003, -0.152706777074, -0.620036176401],
-//     [19.03, -0.133283673636, -0.629186861836],
-//     [19.033, -0.155603343449, -0.646587168999],
-//     [19.06, -0.171737772991, -0.65668353841],
-//     [19.079, -0.190325657536, -0.662944586226],
-//     [19.083000000000002, -0.139584140798, -0.649055978456],
-//     [19.11, -0.158407274796, -0.65153842599],
-//     [19.114, -0.142608862025, -0.681221590377],
-//     [19.142, -0.11737745176, -0.65464803858],
-//     [19.159, -0.133726446229, -0.630357023656],
-//     [19.164, -0.127494249172, -0.597205000369],
-//     [19.192, -0.140456870858, -0.585238126566],
-//     [19.195, -0.143094287755, -0.609122173214],
-//     [19.223, -0.130515544511, -0.601597182461],
-//     [19.240000000000002, -0.0989860504127, -0.603602223617],
-//     [19.255, -0.133790081216, -0.615468201556],
-//     [19.273, -0.130581644476, -0.644080455049],
-//     [19.275000000000002, -0.101359819191, -0.691730385885],
-//     [19.303, -0.0941834242076, -0.723058694018],
-//     [19.32, -0.0402614141598, -0.722327009537],
-//     [19.326, -0.0656051597834, -0.738049299658],
-//     [19.348, -0.033594397154, -0.755812709284],
-//     [19.365000000000002, -0.0401628194648, -0.729644328286],
-//     [19.383, -0.00232569846224, -0.676973186304],
-//     [19.400000000000002, -0.00565461252537, -0.656961685857],
-//     [19.408, -0.0282329468533, -0.626126531819],
-//     [19.435, 0.0224995276422, -0.59165804129],
-//     [19.438, 0.0160159404685, -0.595969282736],
-//     [19.465, 0.0417648118444, -0.585109482125],
-//     [19.483, 0.0221069678926, -0.538953129572],
-//     [19.489, 0.0478040602175, -0.557915454526],
-//     [19.517, 0.0220697139743, -0.5430708641],
-//     [19.519000000000002, 0.0415045484162, -0.565458396819],
-//     [19.547, 0.0283825785775, -0.53791899826],
-//     [19.564, 0.0633351508212, -0.497883424692],
-//     [19.569, 0.0724174463678, -0.480831325454],
-//     [19.597, 0.021447005874, -0.48099890845],
-//     [19.599, 0.0533635813986, -0.498961114752],
-//     [19.627, 0.0856626018748, -0.501345644036],
-//     [19.644000000000002, 0.0154759687524, -0.513146884641],
-//     [19.650000000000002, 0.0317806187448, -0.564877257733],
-//     [19.678, -0.00326806710216, -0.573242266224],
-//     [19.68, 0.0320942366851, -0.612535911673],
-//     [19.709, 0.00999222176255, -0.651205010372],
-//     [19.727, -0.0280591312666, -0.658248745198],
-//     [19.732, -0.0120422302996, -0.664270621913],
-//     [19.759, -0.0443855060977, -0.614931219424],
-//     [19.761, -0.0284196579269, -0.610993928254],
-//     [19.789, -0.0191669856202, -0.575004509192],
-//     [19.807000000000002, -0.0353080060269, -0.562908575157],
-//     [19.813, -0.0292432376492, -0.52659477605],
-//     [19.84, -0.0289372566171, -0.553376644677],
-//     [19.843, -0.0609445963542, -0.556681512786],
-//     [19.87, -0.0195450064465, -0.558041980223],
-//     [19.888, -0.0320108637276, -0.576312395033],
-//     [19.894000000000002, -0.0195450064465, -0.535453370349],
-//     [19.922, -0.0328369985458, -0.486879751845],
-//     [19.925, -0.00996459684002, -0.515529968709],
-//     [19.953, -0.00426492710595, -0.471144475106],
-//     [19.97, -0.0487146222577, -0.497821552153],
-//     [19.975, -0.0649154516559, -0.46995878362],
-//     [20.003, -0.0584123317304, -0.484153915189],
-//     [20.005, -0.0516792376838, -0.510145140071],
-//     [20.033, -0.0899910432375, -0.483146789177],
-//     [20.05, -0.0419727747155, -0.515933052061],
-//     [20.057000000000002, -0.0512890627852, -0.555107385189],
-//     [20.084, -0.038433718148, -0.569499525771],
-//     [20.086000000000002, -0.0383296343956, -0.589103524481],
-//     [20.114, -0.0764586609798, -0.572443453419],
-//     [20.132, -0.0413914946241, -0.570944380369],
-//     [20.138, -0.0830438532435, -0.573329237866],
-//     [20.165, -0.0513780415211, -0.531777320303],
-//     [20.168, -0.045268022464, -0.52097169026],
-//     [20.195, -0.073481109184, -0.536952433981],
-//     [20.199, -0.0610492134144, -0.52292114522],
-//     [20.227, -0.105810844675, -0.50309511742],
-//     [20.244, -0.0671471491471, -0.534954381398],
-//     [20.26, -0.0640499965559, -0.563347340771],
-//     [20.269000000000002, -0.0638450440352, -0.574887839991],
-//     [20.297, -0.124223265237, -0.609736675984],
-//     [20.3, -0.0952015665866, -0.668633402807],
-//     [20.328, -0.129402513562, -0.71832821169],
-//     [20.330000000000002, -0.138972676671, -0.76649053336],
-//     [20.358, -0.0716208689853, -0.782831350065],
-//     [20.36, -0.0749433278972, -0.764997419163],
-//     [20.388, -0.110101297917, -0.767620189586],
-//     [20.405, -0.106754653398, -0.752661973632],
-//     [20.412, -0.107256950976, -0.743443026567],
-//     [20.439, -0.117055811908, -0.723464058739],
-//     [20.442, -0.0980101739219, -0.724196944435],
-//     [20.469, -0.0981445975995, -0.729356415085],
-//     [20.487000000000002, -0.133318439452, -0.728197242293],
-//     [20.493000000000002, -0.0755660578338, -0.748936870871],
-//     [20.52, -0.113408867815, -0.801021299275],
-//     [20.537, -0.141805341417, -0.817062673941],
-//     [20.544, -0.141340664924, -0.826335545237],
-//     [20.572, -0.103240536427, -0.836063034352],
-//     [20.574, -0.11302433898, -0.820732963072],
-//     [20.602, -0.10053790544, -0.779173478378],
-//     [20.604, -0.0945003677013, -0.752216548628],
-//     [20.632, -0.113169495162, -0.777917651019],
-//     [20.649, -0.100624731321, -0.738784320962],
-//     [20.655, -0.122925188799, -0.741675344293],
-//     [20.683, -0.122742448059, -0.771369670649],
-//     [20.685, -0.0911282046983, -0.749766318685],
-//     [20.713, -0.107233152093, -0.725530346212],
-//     [20.73, -0.114098246616, -0.687259302952],
-//     [20.737000000000002, -0.107998615518, -0.67410993362],
-//     [20.764, -0.0984273874661, -0.660362489993],
-//     [20.767, -0.0603173194227, -0.623450844753],
-//     [20.794, -0.0603275579108, -0.606350242862],
-//     [20.798000000000002, -0.00959925020975, -0.56964028865],
-//     [20.815, -0.00636660378832, -0.571635061261],
-//     [20.818, -0.0161919373892, -0.530957019201],
-//     [20.845, -0.0223329172687, -0.55001130361],
-//     [20.848, -0.0221155159494, -0.557768693613],
-//     [20.875, -0.0828416669991, -0.566936240602],
-//     [20.879, -0.018892611813, -0.5771444764],
-//     [20.907, -0.0507236774185, -0.613930490215],
-//     [20.924, -0.066482814349, -0.631547349517],
-//     [20.929000000000002, -0.0790203105064, -0.665004401707],
-//     [20.957, -0.0885776697844, -0.671593522042],
-//     [20.959, -0.0817299202966, -0.713132059815],
-//     [20.987000000000002, -0.0465015719056, -0.756707621582],
-//     [21.004, -0.0559467552843, -0.771964163038],
-//     [21.01, -0.0909292355707, -0.800613548629],
-//     [21.039, -0.0586933958586, -0.814317889365],
-//     [21.04, -0.0678572616004, -0.87243257299],
-//     [21.068, -0.115645248541, -0.896122655216],
-//     [21.085, -0.0964772383456, -0.890043261216],
-//     [21.092, -0.105579126016, -0.940015719755],
-//     [21.12, -0.0961235538275, -0.934807358565],
-//     [21.122, -0.0957950644839, -0.965644384729],
-//     [21.149, -0.0666345109353, -0.994624973283],
-//     [21.167, -0.0607277939275, -0.946529929311],
-//     [21.183, -0.0160389506678, -0.957372637323],
-//     [21.202, -0.0798788364989, -0.938060062784],
-//     [21.204, -0.0446952145359, -0.950027058899],
-//     [21.232, -0.0542589979444, -0.963598908279],
-//     [21.234, -0.0957981306252, -0.978598323662],
-//     [21.252, -0.0448049845423, -0.981171048986],
-//     [21.254, -0.0544133216433, -0.968377764125],
-//     [21.282, -0.0765892375067, -1.00721630737],
-//     [21.285, -0.0251542321312, -1.03618294356],
-//     [21.313, -0.0345288217034, -1.05132524262],
-//     [21.330000000000002, -0.0312606986313, -1.09015896999],
-//     [21.335, -0.0501865635544, -1.07749189035],
-//     [21.373, -0.0404539656201, -1.07384057057],
-//     [21.374, -0.0217958317255, -1.01925234026],
-//     [21.393, 0.0193645536434, -0.997107769331],
-//     [21.411, -0.0286318950929, -0.979814836167],
-//     [21.417, -0.009342231176, -0.994995809152],
-//     [21.444, -0.0253970213811, -1.00707044305],
-//     [21.447, -0.0382075885814, -0.98408209242],
-//     [21.477, 0.0317455849404, -0.970532470385],
-//     [21.488, -0.00946543277001, -0.963169529536],
-//     [21.505, 0.0317911192397, -0.953181730073],
-//     [21.523, 0.0152737840331, -0.878869730812],
-//     [21.528, 0.0568165736527, -0.868888229168],
-//     [21.555, 0.0245292487498, -0.838520055186],
-//     [21.574, -0.0075157207589, -0.837897509141],
-//     [21.578, 0.0182809202351, -0.852846253676],
-//     [21.605, 0.0277339503189, -0.872134735579],
-//     [21.609, 0.00883021274225, -0.88829614423],
-//     [21.63, 0.0150046910395, -0.883685984206],
-//     [21.648, 0.0281699419117, -0.902359957366],
-//     [21.665, -0.00372083828335, -0.930990931273],
-//     [21.683, -0.0225407379885, -0.954454868611],
-//     [21.69, 0.00606545218457, -0.930462442134],
-//     [21.718, -0.00680134488905, -0.948962278916],
-//     [21.72, -0.00340004800931, -0.954215595535],
-//     [21.748, -0.0098390023155, -0.954482207084],
-//     [21.765, -0.0447153385083, -0.987510621566],
-//     [21.772000000000002, -0.0476886844152, -1.00563067123],
-//     [21.799, -0.0319789558596, -0.999783214953],
-//     [21.815, -0.0155653852143, -1.03055516827],
-//     [21.829, -0.0280263628576, -1.06826730979],
-//     [21.844, 0.00745353075396, -1.09173417743],
-//     [21.862000000000002, 0.0268454485846, -1.11590345424],
-//     [21.879, -0.00547770679704, -1.09852825702],
-//     [21.884, -0.0499637312076, -1.09947686993],
-//     [21.913, -0.00875666333011, -1.08645762866],
-//     [21.93, -0.0338940922366, -1.12834489248],
-//     [21.934, 0.0713668103665, -1.10235377808],
-//     [21.962, 0.0457836127802, -1.09981809795],
-//     [21.965, 0.0170269478528, -1.1202401637],
-//     [21.993000000000002, 0.0390083851986, -1.0829606354],
-//     [21.998, 0.0199277566939, -1.08013657558],
-//     [22.005, 0.0424248824308, -1.08330390647],
-//     [22.023, 0.0420262764399, -1.0602551278],
-//     [22.04, 0.0391595553278, -1.05333515907],
-//     [22.047, 0.0553898923605, -1.05116870679],
-//     [22.074, 0.0262728492129, -1.02590496231],
-//     [22.077, 0.0428564481801, -1.08183967763],
-//     [22.105, -0.0215703615037, -1.07029094276],
-//     [22.123, 0.0230642386569, -1.07286145345],
-//     [22.128, -0.00881715772089, -1.05228257992],
-//     [22.155, -0.00939667875121, -1.01641335805],
-//     [22.159, 0.0285771143342, -0.97721917489],
-//     [22.187, 0.0510510597086, -0.967817078619],
-//     [22.204, 0.0126021770083, -0.954883895913],
-//     [22.21, 0.0125594039851, -0.929020779693],
-//     [22.238, 0.00274180586387, -0.929552395459],
-//     [22.240000000000002, 0.0155438177514, -0.905835552781],
-//     [22.268, 0.0117154480874, -0.873288502953],
-//     [22.285, -0.00723645728888, -0.873533497407],
-//     [22.292, -0.0136082737416, -0.867328957891],
-//     [22.319, 0.0117319590248, -0.855782306576],
-//     [22.322, -0.029666505506, -0.882265423368],
-//     [22.349, -0.0520179197127, -0.875577503069],
-//     [22.367, -0.0709184395125, -0.920023720958],
-//     [22.374, -0.064458524274, -0.943282996634],
-//     [22.402, -0.0543167015014, -0.991750965498],
-//     [22.413, -0.0763785306108, -1.03271963759],
-//     [22.431, -0.0666426601132, -1.06942940573],
-//     [22.433, -0.0630673319582, -1.10141244674],
-//     [22.46, -0.0885863226022, -1.10477711889],
-//     [22.478, -0.066181800595, -1.08770372321],
-//     [22.489, -0.0183244016281, -1.08478047912],
-//     [22.511, -0.04400772559, -1.07847385663],
-//     [22.514, -0.047470623067, -1.06226418475],
-//     [22.542, -0.037896859656, -1.04826574708],
-//     [22.559, -0.025237912082, -1.00993378756],
-//     [22.565, -0.0669708319924, -1.00867504501],
-//     [22.593, -0.073269019824, -1.01999070969],
-//     [22.595, -0.0859895841126, -1.04260192041],
-//     [22.623, -0.0215078649883, -1.08096517965],
-//     [22.64, -0.047068032319, -1.08257333281],
-//     [22.647000000000002, -0.0212500556341, -1.11019827699],
-//     [22.674, -0.00845465725446, -1.14253274286],
-//     [22.677, -0.0339636569325, -1.12902648259],
-//     [22.704, -0.0373355804323, -1.09880658851],
-//     [22.718, 0.016828285603, -1.08353132514],
-//     [22.728, -0.0537638494047, -1.03135834464],
-//     [22.755, -0.00620648750015, -1.0247350516],
-//     [22.758, -0.00638329062213, -0.996786929787],
-//     [22.786, -0.00989464600711, -0.962602515211],
-//     [22.803, -0.0225432558485, -0.960003526643],
-//     [22.82, 0.0474034761012, -0.930813492734],
-//     [22.838, 0.0314918464769, -0.918550578928],
-//     [22.839000000000002, 0.0156068848359, -0.93308802808],
-//     [22.867, 0.0571176748656, -0.935235255104],
-//     [22.87, 0.0284262771287, -0.945901531919],
-//     [22.888, 0.0285511052804, -0.953476166468],
-//     [22.89, -0.0259159014856, -0.952090178385],
-//     [22.918, 0.0122344315479, -0.944650401631],
-//     [22.92, -0.0228828405391, -0.932084169618],
-//     [22.948, -0.0553248458602, -0.889187522726],
-//     [22.968, -0.0268080331404, -0.857781836086],
-//     [22.97, -0.0107973943317, -0.855804197727],
-//     [22.999, -0.0588796979786, -0.799195870729],
-//     [23.002, -0.04310817426, -0.781860931451],
-//     [23.029, 0.00781763737695, -0.764265280279],
-//     [23.033, -0.00822599646412, -0.769960223142],
-//     [23.06, 0.00748248141361, -0.725584797071],
-//     [23.078, -0.00214289191764, -0.7195521304],
-//     [23.097, 0.0135506565689, -0.699587143009],
-//     [23.104, -0.0246905253027, -0.710818488003],
-//     [23.122, -0.00879638578259, -0.724177400775],
-//     [23.139, -0.0599381560495, -0.712932748137],
-//     [23.157, -0.0472380178604, -0.708277548711],
-//     [23.174, -0.0759538394414, -0.722208506233],
-//     [23.192, -0.053348177152, -0.734705256995],
-//     [23.195, -0.0694645696423, -0.733568814313],
-//     [23.223, -0.0661536716705, -0.738906160086],
-//     [23.240000000000002, -0.0500980919464, -0.745215569352],
-//     [23.247, -0.014944348422, -0.739064358672],
-//     [23.274, -0.0498613393349, -0.76645824326],
-//     [23.277, -0.017895384643, -0.759014898573],
-//     [23.304000000000002, -0.0213684016617, -0.733876740259],
-//     [23.322, -0.00559242769349, -0.698433554664],
-//     [23.333000000000002, 0.0547571760336, -0.687959824356],
-//     [23.351, -0.00589431708483, -0.689760703957],
-//     [23.358, -0.00617883995283, -0.638241790794],
-//     [23.387, -0.0129507852765, -0.58689077691],
-//     [23.404, 0.0061991064279, -0.602247426136],
-//     [23.408, 0.00323638912148, -0.607651308643],
-//     [23.435, -0.0064742394524, -0.598833458059],
-//     [23.439, 0.00296911536041, -0.602485472359],
-//     [23.467, -0.0290902926208, -0.583333156986],
-//     [23.469, -0.025929265301, -0.574006266623],
-//     [23.497, -0.00376650839837, -0.558066373173],
-//     [23.514, -0.0100267749846, -0.556987716413],
-//     [23.52, -0.019885333497, -0.528390445017],
-//     [23.548000000000002, 0.0148511488039, -0.484243633376],
-//     [23.55, -0.017255931943, -0.462002056212],
-//     [23.578, 0.0113287598153, -0.427357748674],
-//     [23.602, 0.00485864224379, -0.431387332419],
-//     [23.603, -0.0303542030538, -0.414638790548],
-//     [23.629, -0.043215537645, -0.432758651262],
-//     [23.635, -0.00463151278884, -0.440156680398],
-//     [23.653000000000002, -0.020475298572, -0.445561698838],
-//     [23.681, -0.0528407254583, -0.415397467645],
-//     [23.683, -0.0430283451035, -0.428293000984],
-//     [23.711000000000002, -0.0720108899391, -0.39716873914],
-//     [23.713, -0.0493431772796, -0.451941550517],
-//     [23.741, -0.0460230953818, -0.465876854508],
-//     [23.758, -0.103908262414, -0.458579949763],
-//     [23.765, -0.0585318714884, -0.49022662784],
-//     [23.792, -0.10346224202, -0.467464337725],
-//     [23.794, -0.0745486692554, -0.450102201053],
-//     [23.822, -0.0844187664927, -0.422319022351],
-//     [23.839000000000002, -0.0784327245977, -0.39660286458],
-//     [23.845, -0.103922474108, -0.388995070897],
-//     [23.873, -0.0752776342757, -0.391017283226],
-//     [23.877, -0.0590537375483, -0.415964923031],
-//     [23.904, -0.0939144627447, -0.43540728014],
-//     [23.907, -0.1035170634, -0.441246664147],
-//     [23.924, -0.0745953331678, -0.459000387936],
-//     [23.928, -0.116137868126, -0.46461428398],
-//     [23.957, -0.106280492925, -0.484858178099],
-//     [23.958000000000002, -0.12545942479, -0.504316862455],
-//     [23.986, -0.119094442036, -0.485794369036],
-//     [23.988, -0.109918162739, -0.48493601094],
-//     [24.016000000000002, -0.125765122597, -0.482777196556],
-//     [24.033, -0.13887968845, -0.459857423007],
-//     [24.039, -0.10970124288, -0.46133478412],
-//     [24.067, -0.144942131543, -0.435408268135],
-//     [24.071, -0.125648033346, -0.450703162432],
-//     [24.098, -0.1093666313, -0.464058583812],
-//     [24.117, -0.112456323813, -0.48830956359],
-//     [24.121000000000002, -0.137719877464, -0.532934485433],
-//     [24.148, -0.121760865862, -0.535665088908],
-//     [24.152, -0.0989742501987, -0.60242216899],
-//     [24.179000000000002, -0.128051959534, -0.58638051401],
-//     [24.197, -0.101762062972, -0.64930611259],
-//     [24.203, -0.0954363028487, -0.655203682292],
-//     [24.231, -0.111581267969, -0.635736948548],
-//     [24.233, -0.143730721628, -0.606059715288],
-//     [24.261, -0.0636526107763, -0.612079090197],
-//     [24.263, -0.0989742501987, -0.594878295358],
-//     [24.291, -0.127989979544, -0.605279182087],
-//     [24.308, -0.11140983579, -0.612550031113],
-//     [24.314, -0.127680033596, -0.619826353472],
-//     [24.342, -0.0828547411681, -0.630881482458],
-//     [24.344, -0.0633350129676, -0.673998117122],
-//     [24.377, -0.0726701314295, -0.697798926975],
-//     [24.394000000000002, -0.101042026648, -0.74108833032],
-//     [24.396, -0.106866049868, -0.775696160904],
-//     [24.423000000000002, -0.122808496957, -0.770620332949],
-//     [24.427, -0.0686155204987, -0.783581143347],
-//     [24.454, -0.136013862242, -0.757400086407],
-//     [24.472, -0.120082394385, -0.72902276012],
-//     [24.477, -0.110804400223, -0.731712722824],
-//     [24.504, -0.149261528721, -0.739620975481],
-//     [24.507, -0.113938449952, -0.738445309833],
-//     [24.536, -0.116879428586, -0.752952446894],
-//     [24.553, -0.107227193684, -0.759738529389],
-//     [24.562, -0.100315496445, -0.795307715774],
-//     [24.580000000000002, -0.141965690963, -0.77407702634],
-//     [24.598, -0.0975840759059, -0.775155313774],
-//     [24.616, -0.135569106758, -0.775090918539],
-//     [24.633, -0.107463365364, -0.741303649543],
-//     [24.638, -0.110410616478, -0.722359926998],
-//     [24.666, -0.129773256492, -0.706972752364],
-//     [24.669, -0.107684415269, -0.690830411362],
-//     [24.697, -0.0822255320477, -0.682912209855],
-//     [24.714000000000002, -0.111381696352, -0.669379345237],
-//     [24.719, -0.0790645239207, -0.663364955485],
-//     [24.747, -0.0472462893641, -0.673682349866],
-//     [24.751, -0.117542431023, -0.663666847586],
-//     [24.778000000000002, -0.120398307857, -0.683040299297],
-//     [24.796, -0.0979914390178, -0.692871654623],
-//     [24.801000000000002, -0.0916041871905, -0.684687474962],
-//     [24.828, -0.104416979984, -0.670004980602],
-//     [24.832, -0.142679446155, -0.692970397657],
-//     [24.859, -0.0888927820181, -0.638666143378],
-//     [24.872, -0.0887687400076, -0.642457942296],
-//     [24.889, -0.0469510017673, -0.677927064934],
-//     [24.907, -0.11737745176, -0.665954340472],
-//     [24.912, -0.0819071534028, -0.71219501115],
-//     [24.939, -0.081708137718, -0.746337475506],
-//     [24.957, -0.116768344315, -0.744474612804],
-//     [24.963, -0.0846948606951, -0.780216888808],
-//     [24.991, -0.100747611206, -0.785944195256],
-//     [24.993000000000002, -0.10673126906, -0.816438743386],
-//     [25.021, -0.103543111995, -0.821196365766],
-//     [25.038, -0.0526249000157, -0.810003218548],
-//     [25.044, -0.0686225948916, -0.79839362993],
-//     [25.072, -0.0810968053449, -0.822770724026],
-//     [25.074, -0.106353257581, -0.844288571883],
-//     [25.103, -0.0461384250058, -0.80123389831],
-//     [25.121000000000002, -0.0396517286796, -0.813331092295],
-//     [25.126, -0.061781428575, -0.82960726622],
-//     [25.153000000000002, -0.0523871751271, -0.842324081607],
-//     [25.156, -0.0299649307324, -0.85763442504],
-//     [25.183, 0.0152336805354, -0.884816303056],
-//     [25.201, -0.00401155459144, -0.898178966145],
-//     [25.207, -0.0354676547539, -0.905519341182],
-//     [25.234, 0.0216474418862, -0.898273007233],
-//     [25.237000000000002, 0.037776939158, -0.909025286749],
-//     [25.266000000000002, 0.0439761365443, -0.872085491897],
-//     [25.268, 0.0248949206433, -0.866102537331],
-//     [25.288, 0.021696784948, -0.875688326926],
-//     [25.316, 0.0315089969277, -0.883156803916],
-//     [25.319, 0.0505090363748, -0.893685667639],
-//     [25.347, 0.0760194644609, -0.900195353743],
-//     [25.349, 0.0185039644232, -0.8892618155],
-//     [25.377, 0.0693411223931, -0.863244864615],
-//     [25.394000000000002, 0.0469214375613, -0.845808274195],
-//     [25.401, 0.0625610328923, -0.791024267248],
-//     [25.428, 0.0658861419508, -0.813412084896],
-//     [25.431, 0.04303987733, -0.772547413057],
-//     [25.458000000000002, 0.0242627799546, -0.791652296737],
-//     [25.476, 0.0434262896116, -0.797521436839],
-//     [25.481, 0.00799806292648, -0.772533957632],
-//     [25.501, 0.0269029887393, -0.776319632858],
-//     [25.518, 0.0141539060963, -0.732501804996],
-//     [25.536, 0.013923380341, -0.712310519598],
-//     [25.553, 0.0109376916388, -0.720992806591],
-//     [25.572, 0.0136963686839, -0.684293555972],
-//     [25.589000000000002, -0.00876883170409, -0.678661932117],
-//     [25.593, -0.021399722308, -0.693085398176],
-//     [25.621000000000002, -0.0178916627803, -0.733546671222],
-//     [25.638, -0.0206846499975, -0.761311011477],
-//     [25.657, -0.0367942083713, -0.76958289058],
-//     [25.664, -0.0750211631019, -0.758178974041],
-//     [25.682000000000002, -0.0498613393349, -0.755146728887],
-//     [25.699, -0.0558860311265, -0.740335428401],
-//     [25.705000000000002, -0.0462916646169, -0.745668413664],
-//     [25.722, -0.0914487108011, -0.73439461292],
-//     [25.724, -0.0434057984782, -0.724781763824],
-//     [25.752, -0.0463098323786, -0.748015993376],
-//     [25.754, -0.12631638703, -0.764207405122],
-//     [25.782, -0.0751611418956, -0.766421480614],
-//     [25.802, -0.100143750598, -0.79307957016],
-//     [25.806, -0.0748818766603, -0.781051166459],
-//     [25.833000000000002, -0.110060313681, -0.803580603669],
-//     [25.837, -0.131801966339, -0.821371451146],
-//     [25.864, -0.115673151757, -0.814895207846],
-//     [25.867, -0.138351075282, -0.834210677713],
-//     [25.894000000000002, -0.138332347581, -0.806272763782],
-//     [25.908, -0.14527910142, -0.793394731326],
-//     [25.919, -0.135661322862, -0.831267141891],
-//     [25.943, -0.144948357056, -0.871141319656],
-//     [25.948, -0.150960471389, -0.880286866292],
-//     [25.976, -0.121512932362, -0.921118067276],
-//     [25.993000000000002, -0.12785722496, -0.908953403575],
-//     [25.999, -0.114975739289, -0.917976105017],
-//     [26.027, -0.105692263693, -0.877449183012],
-//     [26.031, -0.138106803651, -0.868460677712],
-//     [26.058, -0.0904354371708, -0.829351818446],
-//     [26.061, -0.109688251439, -0.818744547728],
-//     [26.078, -0.116044529717, -0.834635275976],
-//     [26.082, -0.0971326475515, -0.837893706842],
-//     [26.111, -0.119508530206, -0.817705794759],
-//     [26.112000000000002, -0.10284007208, -0.87493120745],
-//     [26.139, -0.0965226673872, -0.850944885961],
-//     [26.142, -0.0837172932899, -0.850451746485],
-//     [26.161, -0.0744480531772, -0.833875874657],
-//     [26.162, -0.0650214100633, -0.824329905738],
-//     [26.189, -0.0680583977291, -0.802481080755],
-//     [26.193, -0.0780239712507, -0.780278037286],
-//     [26.221, -0.0815118819564, -0.748747684221],
-//     [26.223, -0.036804183374, -0.752881994362],
-//     [26.251, -0.0651965720698, -0.772530441146],
-//     [26.268, -0.0905947719483, -0.778045507674],
-//     [26.274, -0.0748352972791, -0.761974407929],
-//     [26.302, -0.12582310493, -0.789466847203],
-//     [26.306, -0.0781088971059, -0.797126104735],
-//     [26.333000000000002, -0.071382128019, -0.803483236853],
-//     [26.351, -0.0966193804166, -0.836647361084],
-//     [26.356, -0.115748190439, -0.871376425973],
-//     [26.383, -0.131253654547, -0.913814534815],
-//     [26.387, -0.0958043603601, -0.942698930148],
-//     [26.416, -0.121254727743, -0.960604643725],
-//     [26.433, -0.101730174589, -0.989335200977],
-//     [26.455000000000002, -0.117805593341, -1.00699616249],
-//     [26.464000000000002, -0.0919268368863, -1.02532778887],
-//     [26.467000000000002, -0.108003423063, -1.00878772135],
-//     [26.494, -0.0765174840869, -0.987132338432],
-//     [26.512, -0.0957516691539, -0.965046617119],
-//     [26.518, -0.0959283588651, -0.950226227715],
-//     [26.547, -0.0800154110104, -0.963018094916],
-//     [26.549, -0.0479367225119, -0.951520788799],
-//     [26.577, -0.0960339038902, -0.962132840463],
-//     [26.579, -0.0800496538576, -0.963614590368],
-//     [26.607, -0.0927116935433, -0.982467774999],
-//     [26.624000000000002, -0.0382498033688, -0.971552177126],
-//     [26.631, -0.0732416410497, -0.960400756003],
-//     [26.658, -0.0609562863566, -0.939253189796],
-//     [26.662, -0.0323996621258, -0.9143224917],
-//     [26.689, -0.0168310207765, -0.889037137467],
-//     [26.707, -0.0457280375158, -0.861012320723],
-//     [26.712, -0.0103671507021, -0.874229094904],
-//     [26.739, -0.0296378407788, -0.871200615864],
-//     [26.743000000000002, -0.032538725772, -0.890229772096],
-//     [26.771, -0.000748703743547, -0.881650952887],
-//     [26.788, -0.0579652455394, -0.894646643272],
-//     [26.793, -0.0418916485282, -0.893529221803],
-//     [26.821, -0.0389429348125, -0.885328840765],
-//     [26.824, -0.048481134359, -0.900329542911],
-//     [26.852, -0.0614259579045, -0.893274424313],
-//     [26.854, -0.0418026413721, -0.911724460075],
-//     [26.882, -0.0641784687326, -0.911694697041],
-//     [26.901, -0.0419149608488, -0.907362795801],
-//     [26.906, -0.0320626914143, -0.934728706203],
-//     [26.937, -0.0196572147799, -0.898110339003],
-//     [26.938, -0.0225739342659, -0.937063270405],
-//     [26.956, -0.0385703006681, -0.938170822285],
-//     [26.983, -0.0319104716041, -0.961450769292],
-//     [26.987000000000002, -0.0256055464059, -0.940842032631],
-//     [27.014, -0.047983199612, -0.962227813258],
-//     [27.018, -0.0380798387711, -0.985736597924],
-//     [27.047, -0.057553607108, -0.960167038739],
-//     [27.064, -0.0510379774637, -0.967442734804],
-//     [27.068, -0.0385816488166, -0.920224761243],
-//     [27.096, -0.0863117377163, -0.951698628368],
-//     [27.098, -0.0705868860693, -0.951750693971],
-//     [27.126, -0.0638248719316, -0.945388521456],
-//     [27.143, -0.0801803835821, -0.968578153155],
-//     [27.151, -0.0831050897053, -0.977600841737],
-//     [27.169, -0.11527042105, -0.937937215408],
-//     [27.187, -0.0834542636084, -0.927572825083],
-//     [27.204, -0.0931256517543, -0.91504330008],
-//     [27.222, -0.0770076253674, -0.886800241852],
-//     [27.239, -0.0520841491446, -0.837010416478],
-//     [27.257, -0.0424869849392, -0.830988488565],
-//     [27.261, -0.0907859869965, -0.818496201198],
-//     [27.288, -0.0649522436188, -0.804739639589],
-//     [27.306, -0.05256644605, -0.779105096146],
-//     [27.312, -0.0750371250931, -0.762669728108],
-//     [27.34, -0.0369102370384, -0.744369398759],
-//     [27.343, -0.0527237147037, -0.746825672065],
-//     [27.371000000000002, -0.0276766212371, -0.724855423711],
-//     [27.373, -0.0208710870215, -0.727792624266],
-//     [27.401, -0.0849678752562, -0.708029815985],
-//     [27.418, -0.120270291334, -0.687272869265],
-//     [27.424, -0.0660654396305, -0.704500525995],
-//     [27.452, -0.079174488596, -0.656011812953],
-//     [27.456, -0.0502563843324, -0.681854171397],
-//     [27.483, -0.0405208505604, -0.681288824681],
-//     [27.501, -0.0438555441128, -0.688555392373],
-//     [27.506, -0.0373885137991, -0.701906384473],
-//     [27.533, -0.0179967381822, -0.730533102832],
-//     [27.537, -0.0622914109637, -0.726917841301],
-//     [27.566, -0.0690829112673, -0.692759801898],
-//     [27.583000000000002, -0.078668142217, -0.710897120179],
-//     [27.587, -0.0658453051603, -0.703410155595],
-//     [27.614, -0.0630617735807, -0.684940156936],
-//     [27.622, -0.0632294464115, -0.677576510212],
-//     [27.638, -0.037428554437, -0.663623806583],
-//     [27.656, -0.0021978135912, -0.660673540028],
-//     [27.658, -0.0346181337026, -0.633053315368],
-//     [27.676000000000002, -0.0694083861707, -0.638354844129],
-//     [27.693, -0.0856872372456, -0.629140646324],
-//     [27.701, -0.0919548699051, -0.639837659837],
-//     [27.728, -0.0825709500838, -0.611662677018],
-//     [27.729, -0.0824206017463, -0.630944675241],
-//     [27.757, -0.0890900960971, -0.600705165515],
-//     [27.774, -0.089287291674, -0.57408120629],
-//     [27.781, -0.127432437825, -0.581731136767],
-//     [27.808, -0.0697410131897, -0.618448829844],
-//     [27.812, -0.101537179872, -0.618571846326],
-//     [27.839000000000002, -0.0856994792412, -0.625162996837],
-//     [27.857, -0.0947919537648, -0.644179993261],
-//     [27.863, -0.110905507992, -0.646932018814],
-//     [27.891000000000002, -0.0882495588517, -0.703365713439],
-//     [27.893, -0.154994144617, -0.703404945431],
-//     [27.921, -0.174170661209, -0.758299596476],
-//     [27.924, -0.199726432281, -0.796266365661],
-//     [27.947, -0.154307240562, -0.797449328128],
-//     [27.965, -0.167006862244, -0.827556094933],
-//     [27.982, -0.224419965162, -0.843819502496],
-//     [28.0, -0.202495072171, -0.827508187973],
-//     [28.017, -0.205028025687, -0.86570949107],
-//     [28.025000000000002, -0.230810640981, -0.873697560439],
-//     [28.052, -0.23357338628, -0.905084667598],
-//     [28.056, -0.227032377849, -0.920792779875],
-//     [28.083000000000002, -0.191883734949, -0.92699753921],
-//     [28.086000000000002, -0.214395219675, -0.929241059251],
-//     [28.113, -0.246355988939, -0.913757789559],
-//     [28.131, -0.213776061339, -0.946698425494],
-//     [28.137, -0.239531991022, -0.955042688144],
-//     [28.165, -0.200799948567, -0.953093239343],
-//     [28.167, -0.200695839417, -0.941218111314],
-//     [28.194, -0.21428294335, -0.936208267247],
-//     [28.217000000000002, -0.204226983521, -0.93368078457],
-//     [28.218, -0.233450511778, -0.909456946676],
-//     [28.246000000000002, -0.176403387716, -0.885137446702],
-//     [28.249000000000002, -0.154176220851, -0.867203802947],
-//     [28.278000000000002, -0.189317380579, -0.857531133326],
-//     [28.296, -0.198902281642, -0.864136477464],
-//     [28.299, -0.211516667756, -0.885371808856],
-//     [28.327, -0.185501687293, -0.876570692389],
-//     [28.331, -0.195230023745, -0.867795805846],
-//     [28.358, -0.182395201796, -0.844142366033],
-//     [28.361, -0.160134975497, -0.845675538373],
-//     [28.388, -0.131806827911, -0.807343461696],
-//     [28.406000000000002, -0.160799785582, -0.764004770811],
-//     [28.412, -0.199622545536, -0.750447044277],
-//     [28.44, -0.190268620468, -0.694653027953],
-//     [28.442, -0.161683972638, -0.70025552253],
-//     [28.47, -0.154910428968, -0.71210689637],
-//     [28.487000000000002, -0.180615949152, -0.69623583822],
-//     [28.493000000000002, -0.161466499474, -0.727392623305],
-//     [28.521, -0.17741477648, -0.717097823833],
-//     [28.525000000000002, -0.152080853772, -0.684096258344],
-//     [28.552, -0.123360722338, -0.684617594154],
-//     [28.572, -0.168161793329, -0.674929829805],
-//     [28.575, -0.155026086511, -0.704314179105],
-//     [28.602, -0.161469040236, -0.703372548512],
-//     [28.606, -0.177128543936, -0.723036157932],
-//     [28.633, -0.129399939663, -0.675799888005],
-//     [28.651, -0.151530111174, -0.720060779229],
-//     [28.656000000000002, -0.164351004748, -0.716248224981],
-//     [28.683, -0.193213542427, -0.754489481526],
-//     [28.687, -0.167299849523, -0.787088893407],
-//     [28.715, -0.154639462249, -0.786855577196],
-//     [28.717000000000002, -0.167234107609, -0.837465017579],
-//     [28.745, -0.214739808474, -0.849657810055],
-//     [28.762, -0.135346667441, -0.827004864638],
-//     [28.768, -0.115730804672, -0.865199037182],
-//     [28.796, -0.205280802804, -0.84387056895],
-//     [28.799, -0.170219200227, -0.838435877304],
-//     [28.827, -0.128352006996, -0.831491963489],
-//     [28.845, -0.11248393951, -0.834864279343],
-//     [28.851, -0.119106622048, -0.803090369353],
-//     [28.873, -0.141887942457, -0.799478928717],
-//     [28.88, -0.122311186915, -0.82015740787],
-//     [28.907, -0.0933499208669, -0.848129430261],
-//     [28.925, -0.0839377619857, -0.829649075262],
-//     [28.931, -0.109186758299, -0.89841809383],
-//     [28.958000000000002, -0.144017827085, -0.901606232428],
-//     [28.962, -0.102358856507, -0.915951156465],
-//     [28.990000000000002, -0.121691301577, -0.907178579426],
-//     [29.007, -0.0928245334137, -0.931923616173],
-//     [29.012, -0.0673681641212, -0.914624091627],
-//     [29.04, -0.0900123290875, -0.881968188477],
-//     [29.043, -0.0994704216516, -0.887199294414],
-//     [29.071, -0.090034277795, -0.866328916683],
-//     [29.073, -0.0550592219412, -0.853078554585],
-//     [29.091, -0.0421763770979, -0.861211046872],
-//     [29.093, -0.0743079005476, -0.871371204702],
-//     [29.121000000000002, -0.0620095044786, -0.814329548262],
-//     [29.124000000000002, -0.045833227054, -0.817626174884],
-//     [29.152, -0.0329257970438, -0.813402136176],
-//     [29.154, -0.0332433555378, -0.787465411317],
-//     [29.175, -0.0555995869917, -0.756840086849],
-//     [29.202, -0.059081893704, -0.722089881999],
-//     [29.206, -0.0369723891107, -0.699087424129],
-//     [29.233, -0.00236610722536, -0.658974096084],
-//     [29.236, -0.0150618135081, -0.664405878167],
-//     [29.264, -0.0118841848987, -0.645657784861],
-//     [29.281000000000002, -0.0503804964987, -0.615364698776],
-//     [29.287, -0.0473704040283, -0.61713294804],
-//     [29.315, -0.0537207429342, -0.640348465142],
-//     [29.318, -0.0251536386806, -0.619559658869],
-//     [29.346, -0.0249648966981, -0.611651638712],
-//     [29.363, -0.0826066636045, -0.608282252427],
-//     [29.37, -0.111309968973, -0.606617783534],
-//     [29.397000000000002, -0.0983972598842, -0.605290342064],
-//     [29.398, -0.0408426945164, -0.613798151834],
-//     [29.426000000000002, -0.0919652736609, -0.617016592817],
-//     [29.443, -0.107802457981, -0.623148340351],
-//     [29.45, -0.0819644137272, -0.6530150545],
-//     [29.477, -0.095087096123, -0.645805502717],
-//     [29.48, -0.146077920184, -0.638340335697],
-//     [29.51, -0.120705585298, -0.634200615024],
-//     [29.527, -0.133928489203, -0.613286549882],
-//     [29.531000000000002, -0.149923210864, -0.595343895278],
-//     [29.558, -0.13089205688, -0.561214139374],
-//     [29.561, -0.105300063444, -0.558069185501],
-//     [29.588, -0.191545590766, -0.528342856988],
-//     [29.606, -0.124537152428, -0.537152437736],
-//     [29.611, -0.178740603388, -0.53251362579],
-//     [29.638, -0.114983466665, -0.524086656151],
-//     [29.642, -0.153389074622, -0.50906506464],
-//     [29.67, -0.140619645194, -0.504996787746],
-//     [29.687, -0.108848135607, -0.470557359582],
-//     [29.692, -0.115659965693, -0.411159529788],
-//     [29.72, -0.12854459137, -0.388234458579],
-//     [29.722, -0.126139039141, -0.357463088317],
-//     [29.75, -0.11643439243, -0.355680245106],
-//     [29.767, -0.113644001921, -0.312486651987],
-//     [29.785, -0.142403823769, -0.310463365079],
-//     [29.794, -0.113416612084, -0.319500543193],
-//     [29.811, -0.155001861736, -0.294246632781],
-//     [29.828, -0.145261099068, -0.318844715768],
-//     [29.846, -0.11936882564, -0.322156195395],
-//     [29.863, -0.132239186582, -0.303178245007],
-//     [29.881, -0.119675295516, -0.307175370164],
-//     [29.885, -0.138898837056, -0.31851453714],
-//     [29.912, -0.113614752868, -0.319130484588],
-//     [29.916, -0.177939551305, -0.331403227919],
-//     [29.933, -0.168426568302, -0.329707835582],
-//     [29.936, -0.110663346768, -0.325986313652],
-//     [29.963, -0.113780240598, -0.331010347573],
-//     [29.967000000000002, -0.148341641824, -0.358354437388],
-//   ]
